% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_echemportal_api_dict.R
\name{build_echa_echemportal_api_dict}
\alias{build_echa_echemportal_api_dict}
\title{code to create ECHA echemportal api dict}
\usage{
build_echa_echemportal_api_dict(toxval.db, filepath)
}
\arguments{
\item{toxval.db}{The version of toxval into which the dictionary is loaded.}

\item{filepath}{The path for all the input xlsx files ./echa_echemportal_api/echa_echemportal_api_files}
}
\description{
code to create ECHA echemportal api dict
}
