% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cas_checkSum.R
\name{cas_checkSum}
\alias{cas_checkSum}
\title{cas_checkSum}
\usage{
cas_checkSum(x)
}
\arguments{
\item{x}{chr. Input vector of values to check. Standard CAS notation using hyphens is fine, as #' all non-digit characters are stripped for checksum calculation. Each element of \emph{x} should contain #' only one suspected CAS RN to check.}
}
\value{
A \code{logical} vector of length \emph{x} denoting whether each \emph{x} is a valid CAS by the checksum method. \code{NA}
input values will be set 0 (FAIL).
}
\description{
Check CAS RN validity via checksum method

For a suspected CAS RN, determine validity by calculating final digit checksum
}
\details{
This function performs a very specific type of check for CAS validity, namely whether the final digit checksum follows
the CAS standard. By default, it also ensures that the digit length is compatible with CAS standards.
}
\examples{
cas_good <- c("71-43-2", "18323-44-9", "7732-18-5") # benzene, clindamycin, water
cas_bad  <- c("61-43-2", "18323-40-9", "7732-18-4") # single digit change from good
cas_checkSum(c(cas_good, cas_bad))
}
\seealso{
\code{\link[stringr]{str_detect}}, \code{\link[stringr]{str_pad}}
}
