% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem.check.R
\name{chem.check}
\alias{chem.check}
\title{FUNCTION_TITLE}
\usage{
chem.check(
  res0,
  name.col = "name",
  casrn.col = "casrn",
  source = NULL,
  verbose = F
)
}
\arguments{
\item{res0}{The data frame in which chemicals names and CASRN will be replaced}

\item{name.col}{The column name that contains the chemical names}

\item{casrn.col}{The column name that contains the CARN values}

\item{source}{The source to be processed. If source=NULL, process all sources}

\item{verbose}{If TRUE, print diagnostic messages}
}
\value{
Return a list with fixed CASRN and name and flags indicating if fixes were made:
res0=res0,name.OK=name.OK,casrn.OK=casrn.OK,checksum.OK=checksum.OK
}
\description{
FUNCTION_DESCRIPTION
}
\details{
#' Check the chemicals from a file
Names with special characters are cleaned and trimmed
CASRN are fixed (dashes put in, trimmed) and check sums are calculated
The output is sent to a file called chemcheck.xlsx in the source data file
One option for using this is to edit the source file until no errors are found

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringi]{stri_escape_unicode}}
 \code{\link[stringr]{str_replace}}, \code{\link[stringr]{str_trim}}
 \code{\link[tidyr]{reexports}}
 \code{\link[openxlsx]{write.xlsx}}
 \code{\link[gsubfn]{list}}
}
