% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.last.character.R
\name{clean.last.character}
\alias{clean.last.character}
\title{clean.last.character}
\usage{
clean.last.character(x)
}
\arguments{
\item{x}{String to be cleaned}
}
\value{
The cleaned string
}
\description{
Clean unneeded characters from the end of a string
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_trim}}
}
