% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.toxval.by.source.R
\name{clean.toxval.by.source}
\alias{clean.toxval.by.source}
\title{Delete a portion of the contents of the toxval database}
\usage{
clean.toxval.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxval from which the data is deleted.}

\item{source}{The data source name}
}
\value{
The database will be altered
}
\description{
Delete a portion of the contents of the toxval database
}
