% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clowder_id_prep.v3.R
\name{clowder_id_prep.v3}
\alias{clowder_id_prep.v3}
\title{Organize the clowder_id and document_name information}
\usage{
clowder_id_prep.v3(db = "dev_toxval_v9")
}
\arguments{
\item{db}{The version of toxval into which the source is loaded.

File from clowder linking clowder_ids to document_names, generated by Taylor Wall
clowder_doc_maps_20220608.xlsx}
}
\description{
Organize the clowder_id and document_name information
}
