% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.pfas_150_sem.to.hawc.R
\name{compare.pfas_150_sem.to.hawc}
\alias{compare.pfas_150_sem.to.hawc}
\title{Compare teh PFAS 150 SEM and hte corresponding HAWC data}
\usage{
compare.pfas_150_sem.to.hawc(db)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}
}
\description{
Compare teh PFAS 150 SEM and hte corresponding HAWC data
}
