% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.fields.to.json.R
\name{convert.fields.to.json}
\alias{convert.fields.to.json}
\title{convert.fields.to.json}
\usage{
convert.fields.to.json(in_dat)
}
\arguments{
\item{in_dat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Combine non-ID columns from audit table into JSON format for audit storage
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{summarise}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{bind}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
}
