% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAT.pipe.source.audit.R
\name{DAT.pipe.source.audit}
\alias{DAT.pipe.source.audit}
\title{Processes DAT QC audit information into database}
\usage{
DAT.pipe.source.audit(source, db, live_df, audit_df)
}
\arguments{
\item{source}{name of ToxVal source table audit information is associated with}

\item{db}{the name of the database}

\item{live_df}{a filepath to the DAT live data to push to the 'source' table}

\item{audit_df}{a filepath to the DAT audit data to push to source_audit}
}
\description{
Processes DAT QC audit information into database
}
