% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_lineage_sync_clowder_metadata.R
\name{doc_lineage_sync_clowder_metadata}
\alias{doc_lineage_sync_clowder_metadata}
\title{doc_lineage_sync_clowder_metadata}
\usage{
doc_lineage_sync_clowder_metadata(
  source_table,
  db,
  clowder_url,
  clowder_api_key,
  batch_size = 250
)
}
\arguments{
\item{source_table}{The source table name (e.g. source_test)}

\item{db}{the name of the database}

\item{clowder_url}{URL to Clowder}

\item{clowder_api_key}{API key to access Clowder resources}

\item{batch_size}{PARAM_DESCRIPTION, Default: 250}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Utility script to sync the Clowder metadata to the database based on Clowder ID
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{add_headers}}, \code{\link[httr]{content}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[tidyr]{nest}}, \code{\link[tidyr]{reexports}}
 \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{bind}}
 \code{\link[purrr]{keep}}
 \code{\link[stringr]{str_replace}}, \code{\link[stringr]{str_trim}}
}
