% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echemportal.prep.v2.step3.R
\name{echemportal.prep.v2.step3}
\alias{echemportal.prep.v2.step3}
\title{Process the ECHA eChemPortal data from 2020 - step 3}
\usage{
echemportal.prep.v2.step3(do.load = F)
}
\arguments{
\item{do.load}{If TRUE, laod all of the in vivo data}
}
\description{
Process the ECHA eChemPortal data from 2020 - step 3
}
