% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecotox.datahub.to.file.R
\name{ecotox.datahub.to.file}
\alias{ecotox.datahub.to.file}
\title{Extract ECOTOX from the datahub to a file}
\usage{
ecotox.datahub.to.file(toxval.db, verbose = T, do.load = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{verbose}{Whether the loaded rows should be printed to the console.}

\item{do.load}{If TRUE, load the data from the input file and put into a global variable}
}
\description{
Extract ECOTOX from the datahub to a file
}
