% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efsa.clean.R
\name{efsa.clean}
\alias{efsa.clean}
\title{prepare EFSA_combined.xlsx using dictionaries from efsa.dict.prep.R to be 
loaded into dev_toxval_source_v4.}
\usage{
efsa.clean(dir, verbose = F)
}
\arguments{
\item{dir}{The directory where the input data sits, ../efsa2/efsa2_files/}
}
\description{
prepare EFSA_combined.xlsx using dictionaries from efsa.dict.prep.R to be 
loaded into dev_toxval_source_v4.
}
