% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efsa.dict.prep.R
\name{efsa.dict.prep}
\alias{efsa.dict.prep}
\title{Create dictionaries for preparing EFSA_combined.xlsx produced from efsa.prep.R}
\usage{
efsa.dict.prep(dir)
}
\arguments{
\item{dir}{The directory where the input data sits, ../efsa2/efsa2_files/}
}
\description{
Create dictionaries for preparing EFSA_combined.xlsx produced from efsa.prep.R
}
