% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efsa.prep.R
\name{efsa.prep}
\alias{efsa.prep}
\title{Prepare the new EFSA data}
\usage{
efsa.prep(dir, step1 = F, step2 = F, step3 = F, step4 = F)
}
\arguments{
\item{dir}{The directory where the input data sits, ../efsa2/efsa2_files/}

\item{step1}{Run the first step of the processing}

\item{step2}{Run the second step of the processing}

\item{step3}{Run the third step of the processing}

\item{step4}{Run the fourth step of the processing}
}
\description{
Prepare the new EFSA data
}
