% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.all.by.source.for.release.R
\name{export.all.by.source.for.release}
\alias{export.all.by.source.for.release}
\title{Build a data frame of the data from toxval and export by source as a
series of xlsx files. This only has the columns tht someone outside might use}
\usage{
export.all.by.source.for.release(toxval.db, source = NULL)
}
\arguments{
\item{toxval.db}{Database version}

\item{source}{The source to be updated
#' @return for each source writes an Excel file with the name
 ../export/export_by_source_{data}/toxval_all_{toxval.db}_{source}.xlsx}
}
\description{
Build a data frame of the data from toxval and export by source as a
series of xlsx files. This only has the columns tht someone outside might use
}
