% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.all.by.source.pfas.R
\name{export.all.by.source.pfas}
\alias{export.all.by.source.pfas}
\title{Build a data frame of the data from toxval and export by source as a
series of xlsx files. THis is a version for PFAS specific use}
\usage{
export.all.by.source.pfas(toxval.db)
}
\arguments{
\item{toxval.db}{Database version}

\item{source}{The source to be updated}
}
\value{
Write a file with the results
}
\description{
Build a data frame of the data from toxval and export by source as a
series of xlsx files. THis is a version for PFAS specific use
}
