% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.all.with.references v92.R
\name{export.all.with.references.v92}
\alias{export.all.with.references.v92}
\title{Build a data frame of the PODs and exports as xlsx}
\usage{
export.all.with.references.v92(toxval.db = "res_toxval_v92", file.name = NA)
}
\arguments{
\item{toxval.db}{Database version}

\item{file.name}{If not NA, this is a file containing chemicals, and only those chemicals will be exported}

\item{human_eco}{Either 'human health' or 'eco'}
}
\value{
writes an Excel file with the name
 ../export/toxval_pod_summary_[human_eco]_Sys.Date().xlsx
}
\description{
Build a data frame of the PODs and exports as xlsx
}
