% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.bcfbaf.R
\name{export.bcfbaf}
\alias{export.bcfbaf}
\title{Export the BCF / BAF data data}
\usage{
export.bcfbaf(toxval.db, dir = "./export")
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes and Excel file with the name
 ../qc_export/toxval_genetox_details_[Sys.Date].xlsx
}
\description{
Export the BCF / BAF data data
}
