% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.cancer.studies.R
\name{export.cancer.studies}
\alias{export.cancer.studies}
\title{Build a data frame of the cancer experimental records}
\usage{
export.cancer.studies(toxval.db)
}
\arguments{
\item{toxval.db}{Database version}

\item{source}{The source to be updated}
}
\value{
Write a file with the results
}
\description{
Build a data frame of the cancer experimental records
}
