% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.cancer.summary.R
\name{export.cancer.summary}
\alias{export.cancer.summary}
\title{Build a data frame of the Cancer calls and exports as xlsx}
\usage{
export.cancer.summary(toxval.db, file.name = NA, dir = "./export")
}
\arguments{
\item{toxval.db}{Database version}

\item{file.name}{If not NULL, read afles of casrn from the chemical folder and only export
those chemicals}
}
\value{
writes and Excel file with the name
 ../qc_export/toxval_pod_summary_min_quality_id_[human_eco]_[min_quality_id]_[Sys.Date].xlsx
}
\description{
Build a data frame of the Cancer calls and exports as xlsx
}
