% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.cgse.samples.R
\name{export.cgse.samples}
\alias{export.cgse.samples}
\title{Export samples of the cancer, genetox, and skin/eye data}
\usage{
export.cgse.samples(toxval.db = "res_toxval_v93")
}
\arguments{
\item{toxval.db}{Database version}
}
\description{
Export samples of the cancer, genetox, and skin/eye data
}
