% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.chemicals.R
\name{export.chemicals}
\alias{export.chemicals}
\title{Export all chemicals in the chemical and chemical_list tables}
\usage{
export.chemicals(toxval.db)
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes and Excel file with the name
 ../export/toxval_chemicals_[Sys.Date].xlsx
}
\description{
Export all chemicals in the chemical and chemical_list tables
}
