% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.clowder.status.R
\name{export.clowder.status}
\alias{export.clowder.status}
\title{Export the table of the sources for the manuscript
#' @param toxval.db Database version}
\usage{
export.clowder.status(toxval.db, source.db)
}
\description{
Export the table of the sources for the manuscript
#' @param toxval.db Database version
}
