% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.echa.repeats.R
\name{export.echa.repeats}
\alias{export.echa.repeats}
\title{Export ECHA data from all four ECHA sources and look for overlaps}
\usage{
export.echa.repeats(toxval.db)
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes and Excel file with the name
 ./export/toxval_echa_repeats_[Sys.Date].xlsx
}
\description{
Export ECHA data from all four ECHA sources and look for overlaps
}
