% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.final.params.R
\name{export.final.params}
\alias{export.final.params}
\title{Export the final values for the character params (e.g. toxval_type).}
\usage{
export.final.params(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\description{
Export the final values for the character params (e.g. toxval_type).
}
