% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.study_type.R
\name{export.for.study_type}
\alias{export.for.study_type}
\title{Export data required for setting the study type}
\usage{
export.for.study_type(toxval.db = "res_toxval_v93")
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes an Excel file with the name
 ../export/toxval_pod_summary_[human_eco]_Sys.Date().xlsx
}
\description{
Export data required for setting the study type
}
