% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.for.toxvaldb.manuscript.R
\name{export.for.toxvaldb.manuscript}
\alias{export.for.toxvaldb.manuscript}
\title{Build a data frame of the data for the toxval manuscript}
\usage{
export.for.toxvaldb.manuscript(toxval.db)
}
\arguments{
\item{toxval.db}{Database version}

\item{source}{The source to be updated}
}
\value{
Write a file with the results
}
\description{
Build a data frame of the data for the toxval manuscript
}
