% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.genetox_details.R
\name{export.genetox_details}
\alias{export.genetox_details}
\title{Export the detailed genetox data}
\usage{
export.genetox_details(toxval.db, dir = "./export")
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes and Excel file with the name
 ../qc_export/toxval_genetox_details_[Sys.Date].xlsx
}
\description{
Export the detailed genetox data
}
