% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.genetox_summary.R
\name{export.genetox_summary}
\alias{export.genetox_summary}
\title{Export the summary genetox data}
\usage{
export.genetox_summary(toxval.db, dir = "./export")
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes and Excel file with the name
 ../qc_export/toxval_genetox_summary_[Sys.Date].xlsx
}
\description{
Export the summary genetox data
}
