% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.missing.rac.by.source.R
\name{export.missing.rac.by.source}
\alias{export.missing.rac.by.source}
\title{Export the rows with a missing risk_assessment_class}
\usage{
export.missing.rac.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{Database version}

\item{source}{The source to be processed}
}
\value{
writes an Excel file with the name
 ./qc_export/toxval_missing_risk_assessment_class_Sys.Date().xlsx"
}
\description{
Export the rows with a missing risk_assessment_class
}
