% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.missing.rac.R
\name{export.missing.rac}
\alias{export.missing.rac}
\title{Export the rows with a missing risk_assessment_class}
\usage{
export.missing.rac(toxval.db)
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes an Excel file with the name
 ./qc_export/toxval_missing_risk_assessment_class_Sys.Date().xlsx"
}
\description{
Export the rows with a missing risk_assessment_class
}
