% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.missing.strains.R
\name{export.missing.strains}
\alias{export.missing.strains}
\title{Set the strain information in toxval}
\usage{
export.missing.strains(toxval.db, date_string = "2022-05-25")
}
\arguments{
\item{toxval.db}{The version of the database to use}

\item{date_string}{The date of the latest dictinary version}
}
\description{
Set the strain information in toxval
}
