% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.pod.summary.R
\name{export.pod.summary}
\alias{export.pod.summary}
\title{Build a data frame of the PODs and exports as xlsx}
\usage{
export.pod.summary(
  toxval.db,
  human_eco = "human health",
  file.name = NA,
  dir = "./export"
)
}
\arguments{
\item{toxval.db}{Database version}

\item{human_eco}{Either 'human health' or 'eco'}

\item{file.name}{If not NA, this is a file containing chemicals, and only those chemicals will be exported}
}
\value{
writes an Excel file with the name
 ../export/toxval_pod_summary_[human_eco]_Sys.Date().xlsx
}
\description{
Build a data frame of the PODs and exports as xlsx
}
