% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.record_source.R
\name{export.record_source}
\alias{export.record_source}
\title{Build a data frame of the th data from record_source and export}
\usage{
export.record_source(toxval.db, dir = "./export")
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes an Excel file with the name
 ../export/toxval_record_summary_[Sys.Date].xlsx
}
\description{
Build a data frame of the th data from record_source and export
}
