% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.skin_eye.R
\name{export.skin_eye}
\alias{export.skin_eye}
\title{Export the skin and eye data}
\usage{
export.skin_eye(toxval.db, dir = "./export")
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes and Excel file with the name
 ../qc_export/toxval_skin_eye_[Sys.Date].xlsx
}
\description{
Export the skin and eye data
}
