% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.source_chemical.R
\name{export.source_chemical}
\alias{export.source_chemical}
\title{Export the source chemical table}
\usage{
export.source_chemical(db)
}
\arguments{
\item{db}{The name of the database String to be cleaned}
}
\description{
Export the source chemical table
}
