% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.toxval_dictionary.R
\name{export.toxval_dictionary}
\alias{export.toxval_dictionary}
\title{Export the toxval_dictionary table}
\usage{
export.toxval_dictionary(toxval.db, dir = "./export")
}
\arguments{
\item{toxval.db}{Database version}
}
\value{
writes and Excel file with the name
 ../qc_export/toxval_dictionary_[Sys.Date].xlsx
}
\description{
Export the toxval_dictionary table
}
