% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_chemicals_to_curate.R
\name{export_chemicals_to_curate}
\alias{export_chemicals_to_curate}
\title{export_chemicals_to_curate}
\usage{
export_chemicals_to_curate(db, export_all = FALSE)
}
\arguments{
\item{db}{Version of toxval_source to use}

\item{export_all}{Whether to export all chemicals, Default: FALSE}
}
\value{
None
}
\description{
Export XLSX files by source for chemical curation
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[separate][tidyr::separate]
 [bind_rows][dplyr::bind_rows], [filter][dplyr::filter], [select][dplyr::select], [mutate][dplyr::mutate], [rowwise][dplyr::rowwise], [case_when][dplyr::case_when], [ungroup][dplyr::ungroup], [group_split][dplyr::group_split]
 [read_xlsx][readxl::read_xlsx]
 [write_xlsx][writexl::write_xlsx]
}
