% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.chemical.R
\name{fill.chemical}
\alias{fill.chemical}
\title{Fill the chemical table}
\usage{
\method{fill}{chemical}(toxval.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Fill the chemical table
}
