% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.chemical_source_index.R
\name{fill.chemical_source_index}
\alias{fill.chemical_source_index}
\title{Load the chemical_source_index table.}
\usage{
\method{fill}{chemical_source_index}(db)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}
}
\description{
Load the chemical_source_index table.
}
