% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.toxval.defaults.global.R
\name{fill.toxval.defaults.global}
\alias{fill.toxval.defaults.global}
\title{Set Toxval Defaults globally}
\usage{
\method{fill}{toxval.defaults.global}(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval from which to set defaults.}
}
\description{
Set Toxval Defaults globally
}
