% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.toxval.defaults.R
\name{fill.toxval.defaults}
\alias{fill.toxval.defaults}
\title{Set Toxval Defaults}
\usage{
\method{fill}{toxval.defaults}(toxval.db, mat)
}
\arguments{
\item{toxval.db}{The version of toxval from which to set defaults.}

\item{mat}{An input matrix of data}
}
\value{
The data matrix afer fixing
}
\description{
Set Toxval Defaults
}
