% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.all.param.by.source.R
\name{fix.all.param.by.source}
\alias{fix.all.param.by.source}
\title{Alter the contents of toxval according to an excel dictionary file with fields -
exposure_method, exposure_route, sex,strain, study_duration_class, study_duration_units, study_type,
toxval_type, exposure_form, media, toxval_subtype}
\usage{
fix.all.param.by.source(toxval.db, source = NULL, fill.toxval_fix = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{source}{The source to be fixed. If source=NULL, fix all sources}

\item{fill.toxval_fix}{If TRUE (default) read the dictionaries into the toxval_fix table}
}
\value{
The database will be altered
}
\description{
Alter the contents of toxval according to an excel dictionary file with fields -
exposure_method, exposure_route, sex,strain, study_duration_class, study_duration_units, study_type,
toxval_type, exposure_form, media, toxval_subtype
}
