% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.all.param.new.by.source.R
\name{fix.all.param.new.by.source}
\alias{fix.all.param.new.by.source}
\title{Alter the contents of toxval according to an excel dictionary file with fields -
exposure_method, exposure_route, sex,strain, study_duration_class, study_duration_units, study_type,
toxval_type, exposure_form, media, toxval_subtype}
\usage{
fix.all.param.new.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Alter the contents of toxval according to an excel dictionary file with fields -
exposure_method, exposure_route, sex,strain, study_duration_class, study_duration_units, study_type,
toxval_type, exposure_form, media, toxval_subtype
}
