% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.casrn.R
\name{fix.casrn}
\alias{fix.casrn}
\title{Fix a CASRN that has one of several problems}
\usage{
fix.casrn(casrn, cname = "", verbose = F)
}
\arguments{
\item{casrn}{Input CASRN to be fixed}

\item{cname}{An optional chemical name}

\item{verbose}{if TRUE, print hte input values}
}
\value{
the fixed CASRN
}
\description{
Fix a CASRN that has one of several problems
}
