% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.clowder_id.by.source.R
\name{fix.clowder_id.by.source}
\alias{fix.clowder_id.by.source}
\title{Add clowder_id's to source db tables based on matching values with source table}
\usage{
fix.clowder_id.by.source(source.db, source_table, source)
}
\arguments{
\item{source.db}{The version of toxval source in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Add clowder_id's to source db tables based on matching values with source table
}
