% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.clowder_id.by.toxval.R
\name{fix.clowder_id.by.toxval}
\alias{fix.clowder_id.by.toxval}
\title{Add clowder_id's to source db tables}
\usage{
fix.clowder_id.by.toxval(toxval.db, source.db, source_table, source, infile)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{infile}{The input file ./clowder_id_mapping/source_doc_with_clowder_id.xlsx, new_qa_set_with_ClowderID_12-22.csv,clowderid_record_source_hash_hawc_pfas_150_430_with_source_hash.xlsx,
atsdr_toxval_record_source_w_document-Aswani.csv,efsa2_toxval_record_source_w_document-Aswani.csv, original_oppt_table_w_document-Aswani.csv,
subset_v8_chiu_record_source_hash-Aswani.csv,WHOIPCS_v8_v9_should_be_same_record_source_hash-Aswani.csv,missing_toxval_source_clowder_corrected.xlsx, toxval_new_qa_set_ClowderID_02-02-ASWANI.csv,
del_docs_matched_20220323.xlsx, toxval_qa_set_2_ECHA-IUCLID_ClowderID_03-23_Aswani.csv}
}
\value{
The database will be altered
}
\description{
Add clowder_id's to source db tables
}
