% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.critical_effect.icf.by.source.R
\name{fix.critical_effect.icf.by.source}
\alias{fix.critical_effect.icf.by.source}
\title{standardize critical_effect in toxval table based on icf dictionary and toxval critical effects dictionary}
\usage{
fix.critical_effect.icf.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{source}{THe source to be fixed}
}
\description{
standardize critical_effect in toxval table based on icf dictionary and toxval critical effects dictionary
}
