% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.critical_effect_icf.R
\name{fix.critical_effect.icf}
\alias{fix.critical_effect.icf}
\title{standardize critical_effect in toxval table based on icf dictionary
Fix occurances of multiple critical_effect in the toxval_critical_effect table based on values from critical_effect dictionary file}
\usage{
fix.critical_effect.icf(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}
}
\description{
standardize critical_effect in toxval table based on icf dictionary
Fix occurances of multiple critical_effect in the toxval_critical_effect table based on values from critical_effect dictionary file
}
