% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.critical_effect.R
\name{fix.critical_effect}
\alias{fix.critical_effect}
\title{Fix critical_effect in toxval table
Fix occurances of multiple critical_effect in the toxval_critical_effect table based on values from critical_effect dictionary file}
\usage{
fix.critical_effect(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}
}
\description{
Fix critical_effect in toxval table
Fix occurances of multiple critical_effect in the toxval_critical_effect table based on values from critical_effect dictionary file
}
