% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.empty.by.source.R
\name{fix.empty.by.source}
\alias{fix.empty.by.source}
\title{Set all empty cells in toxval to '-'}
\usage{
fix.empty.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{source}{The source to be fixed}
}
\value{
The database will be altered
}
\description{
Set all empty cells in toxval to '-'
}
