% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.empty.R
\name{fix.empty}
\alias{fix.empty}
\title{Set all empty cells in toxval to '-'}
\usage{
fix.empty(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Set all empty cells in toxval to '-'
}
