% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.empty.record_source.R
\name{fix.empty.record_source}
\alias{fix.empty.record_source}
\title{Set all empty cells in record_source to '-'}
\usage{
fix.empty.record_source(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Set all empty cells in record_source to '-'
}
