% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.exposure_form.by.source.R
\name{fix.exposure_form.by.source}
\alias{fix.exposure_form.by.source}
\title{Exposure Method temporary fix to add Exposure Form}
\usage{
fix.exposure_form.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The database version to use}
}
\description{
Exposure Method temporary fix to add Exposure Form
}
