% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.exposure_form.R
\name{fix.exposure_form}
\alias{fix.exposure_form}
\title{Exposure Method temporary fix to add Exposure Form}
\usage{
fix.exposure_form(toxval.db)
}
\arguments{
\item{toxval.db}{The database version to use}
}
\description{
Exposure Method temporary fix to add Exposure Form
}
