% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.exposure_route.by.type.new.by.source.R
\name{fix.exposure_route.by.type.new.by.source}
\alias{fix.exposure_route.by.type.new.by.source}
\title{Alter the exposure route and study_duration_class of toxval based on toxval_type}
\usage{
fix.exposure_route.by.type.new.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Alter the exposure route and study_duration_class of toxval based on toxval_type
}
