% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.exposure_route.by.type.R
\name{fix.exposure_route.by.type}
\alias{fix.exposure_route.by.type}
\title{Alter the exposure route of toxval according to an excel dictionary}
\usage{
fix.exposure_route.by.type(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Alter the exposure route of toxval according to an excel dictionary
}
