% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.generation.by.source.R
\name{fix.generation.by.source}
\alias{fix.generation.by.source}
\title{Alter the contents of toxval according to an excel dictionary file with field generation}
\usage{
fix.generation.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{source}{The source to be processes}
}
\value{
The database will be altered
}
\description{
Alter the contents of toxval according to an excel dictionary file with field generation
}
