% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.greek.symbols.R
\name{fix.greek.symbols}
\alias{fix.greek.symbols}
\title{fix.greek.symbols}
\usage{
fix.greek.symbols(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
Returns a dataframe with all character fields edited to replace Greek symbols
with alphabetic equivalents.
}
\description{
A function to check all character fields and replace Greek symbols
with alphabetic equivalents.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
