% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.human_eco.by.source.R
\name{fix.human_eco.by.source}
\alias{fix.human_eco.by.source}
\title{Fix the human_eco flag}
\usage{
fix.human_eco.by.source(toxval.db, source, reset = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Fix the human_eco flag
}
