% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.hyphen.by.source.R
\name{fix.hyphen.by.source}
\alias{fix.hyphen.by.source}
\title{Set all hyphen cells in toxval to 'Not Specified'}
\usage{
fix.hyphen.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Set all hyphen cells in toxval to 'Not Specified'
}
