% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.non_ascii.v2.R
\name{fix.non_ascii.v2}
\alias{fix.non_ascii.v2}
\title{Flag and fix non-ascii characters in the database}
\usage{
fix.non_ascii.v2(df, source)
}
\arguments{
\item{df}{The dataframe to be processed}

\item{The}{source to be fixed}
}
\value{
The dataframe with non ascii characters replaced with cleaned versions
}
\description{
Flag and fix non-ascii characters in the database
}
