% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.priority_id.R
\name{fix.priority_id}
\alias{fix.priority_id}
\title{Fix the priority_id in the toxval table based on source}
\usage{
fix.priority_id(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}
}
\description{
Fix the priority_id in the toxval table based on source
}
