% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.qa_status.by.source.R
\name{fix.qa_status.by.source}
\alias{fix.qa_status.by.source}
\title{Fix the qa_status flag}
\usage{
fix.qa_status.by.source(toxval.db, source, reset = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Fix the qa_status flag
}
