% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.qc_status.by.source.R
\name{fix.qc_status.by.source}
\alias{fix.qc_status.by.source}
\title{Fix the qa_status flag}
\usage{
fix.qc_status.by.source(toxval.db, source = NULL, reset = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{source}{The source to be fixed}

\item{reset}{If TRUE, reset all values to 'pass' before setting}
}
\value{
The database will be altered
}
\description{
Fix the qa_status flag
}
