% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.risk_assessment_class.all.source.R
\name{fix.risk_assessment_class.all.source}
\alias{fix.risk_assessment_class.all.source}
\title{Fix the risk assessment class for all source.}
\usage{
fix.risk_assessment_class.all.source(toxval.db, restart = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{restart}{If TRUE, delete all values and start from scratch}
}
\description{
Fix the risk assessment class for all source.
}
