% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.risk_assessment_class.by.source.R
\name{fix.risk_assessment_class.by.source}
\alias{fix.risk_assessment_class.by.source}
\title{Set the risk assessment class of toxval according to an excel dictionary.
Values may beset multiple times, so the excel sheet should be ordered so that
the last ones to be set are last}
\usage{
fix.risk_assessment_class.by.source(toxval.db, source, restart = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{source}{The source to be updated}

\item{restart}{If TRUE, delete all values and start from scratch}
}
\description{
Set the risk assessment class of toxval according to an excel dictionary.
Values may beset multiple times, so the excel sheet should be ordered so that
the last ones to be set are last
}
