% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.single.param.by.source.R
\name{fix.single.param.by.source}
\alias{fix.single.param.by.source}
\title{Alter the contents of toxval according to an excel dictionary}
\usage{
fix.single.param.by.source(toxval.db, param, source, ignore = FALSE)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{param}{The parameter value to be fixed}

\item{source}{The source to be fixed}

\item{ignore}{If TRUE allow missing values to be ignored}
}
\value{
The database will be altered
}
\description{
Alter the contents of toxval according to an excel dictionary
}
