% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.single.param.new.R
\name{fix.single.param.new}
\alias{fix.single.param.new}
\title{Alter the contents of toxval according to an excel dictionary}
\usage{
fix.single.param.new(toxval.db, param, ignore = FALSE)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{param}{THe parameter value to be fixed}

\item{ignore}{If TRUE allow missing values to be ignored}
}
\value{
The database will be altered
}
\description{
Alter the contents of toxval according to an excel dictionary
}
