% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.source_chemical.name.R
\name{fix.source_chemical.name}
\alias{fix.source_chemical.name}
\title{Set all source_chemical names to the DSSTox preferred_name}
\usage{
fix.source_chemical.name(toxval.db, source = NULL)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{soruce}{The name of the source the be fixed}
}
\value{
The database will be altered
}
\description{
Set all source_chemical names to the DSSTox preferred_name
}
