% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.species_ecotox_by_source.R
\name{fix.species.ecotox.by.source}
\alias{fix.species.ecotox.by.source}
\title{Update the species_id in toxval from species_ecotox.}
\usage{
fix.species.ecotox.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of the database to use}
}
\description{
Update the species_id in toxval from species_ecotox.
}
