% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.species.R
\name{fix.species}
\alias{fix.species}
\title{Update the species information. This method has to be run iteratively to find
the remaining species_original names that don't match what is in the database}
\usage{
fix.species(toxval.db)
}
\arguments{
\item{toxval.db}{The version of the database to use}
}
\description{
Update the species information. This method has to be run iteratively to find
the remaining species_original names that don't match what is in the database
}
