% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.species.v2.R
\name{fix.species.v2}
\alias{fix.species.v2}
\title{Set the species_id column in toxval}
\usage{
fix.species.v2(toxval.db, source, date_string = "2022-05-25")
}
\arguments{
\item{toxval.db}{The version of the database to use}

\item{source}{The source to be fixed}

\item{date_string}{The date version of the dictionary}
}
\description{
This function replaces fix.species
This function precedes toxvaldb.load.species
}
