% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.species_ecotox.R
\name{fix.species_ecotox}
\alias{fix.species_ecotox}
\title{Update the species ecotox table with habitat information from ecotox dictionary.}
\usage{
fix.species_ecotox(toxval.db)
}
\arguments{
\item{toxval.db}{The version of the database to use}
}
\description{
Update the species ecotox table with habitat information from ecotox dictionary.
}
