% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.strain.by.source.R
\name{fix.strain.by.source}
\alias{fix.strain.by.source}
\title{Update the strain_group in toxval from strain_dictionary_2022-03-07.xlsx.}
\usage{
fix.strain.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of the database to use}
}
\description{
Update the strain_group in toxval from strain_dictionary_2022-03-07.xlsx.
}
