% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.strain.v2.R
\name{fix.strain.v2}
\alias{fix.strain.v2}
\title{Set the strain information in toxval}
\usage{
fix.strain.v2(toxval.db, source, date_string = "2022-05-25")
}
\arguments{
\item{toxval.db}{The version of the database to use}

\item{source}{The source to be fixed}

\item{date_string}{The date ofhte latest dictinary version}
}
\description{
Set the strain information in toxval
}
