% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.study_type.manual.R
\name{fix.study_type.manual}
\alias{fix.study_type.manual}
\title{Fix the study_type using manual curation}
\usage{
fix.study_type.manual(toxval.db, source = NULL, sys.date = "2022-10-27")
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Fix the study_type using manual curation
}
