% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.toxval_numeric_qualifier.by.source.R
\name{fix.toxval_numeric_qualifier.by.source}
\alias{fix.toxval_numeric_qualifier.by.source}
\title{Fix the toxval_numeric_qualifier by source}
\usage{
fix.toxval_numeric_qualifier.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Fix the toxval_numeric_qualifier by source
}
