% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.toxval_numeric_qualifier.R
\name{fix.toxval_numeric_qualifier}
\alias{fix.toxval_numeric_qualifier}
\title{Fix the toxval_numeric_qualifier}
\usage{
fix.toxval_numeric_qualifier(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Fix the toxval_numeric_qualifier
}
