% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.units.R
\name{fix.units}
\alias{fix.units}
\title{Do all of the fixes to units}
\usage{
fix.units(toxval.db, do.convert = T)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{do.convert}{If TRUE, so unit conversions, as opposed to just cleaning}
}
\description{
\enumerate{
  \item All of these steps operate on the toxval_units column.
  \item To allow this to be run multiple times during debugging, the first step
  is to copy toxval_units_original into toxval_units
  \item Replace variant unit names with standard ones, running fix.single.param.
  This fixes issues like variant names for mg/kg-day and uses the dictionary
  file dictionary/toxval_units_5.xlsx
  \item Fix special characters in toxval_units
  \item Fix issues with units containing extra characters for some ECOTOX records
  \item Convert units that are multiples of standard ones (e.g. ppb to ppm). This
  uses the dictionary file dictionary/toxval_units conversions 2018-09-12.xlsx
  \item Run conversions from molar to mg units, using MW. This uses the dictionary file
   dictionary/MW conversions.xlsx
  \item Convert ppm to mg/m3 for inhalation studies. This uses the conversion Concentration
  (mg/m3) = 0.0409 x concentration (ppm) x molecular weight. See
  https://cfpub.epa.gov/ncer_abstracts/index.cfm/fuseaction/display.files/fileID/14285.
  This function requires htat the DSSTox external chemical_id be set
  \item Convert ppm to mg/kg-day in toxval according to a species-specific
  conversion factor for oral exposures. This uses the dictionary file
  dictionary/ppm to mgkgday by animal.xlsx
  See: www10.plala.or.jp/biostatistics/1-3.doc
  This probbaly assumes feed rather than water
  \item Make sure that eco studies are in mg/L and human health in mg/m3
}
}
