% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.units.testmw.R
\name{fix.units.testmw}
\alias{fix.units.testmw}
\title{Do all of the fixes to units}
\usage{
fix.units.testmw(toxval.db = "dev_toxval_v9", do.convert = T)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{do.convert}{If TRUE, so unit conversions, as opposed to just cleaning}
}
\description{
Do all of the fixes to units
}
