% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_numeric_unit_split.R
\name{fix_numeric_units_split}
\alias{fix_numeric_units_split}
\title{fix_numeric_units_split}
\usage{
fix_numeric_units_split(
  df,
  to_split = "",
  value_to = "value",
  units_to = "units"
)
}
\arguments{
\item{df}{Input DataFrame}

\item{to_split}{String of the field containing the values to split, Default: ''}

\item{value_to}{String of the name of the field for the split values, Default: 'value'}

\item{units_to}{String of the name of the field for the split units, Default: 'units'}
}
\value{
DataFrame with the split columns based on the input parameters
}
\description{
Generic function to split an input column into value and units columns.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{arrange}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_extract}}, \code{\link[stringr]{str_replace}}
 \code{\link[rlang]{sym}}
 \code{\link[tidyr]{separate}}
}
