% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rsl_excel.R
\name{format_rsl_excel}
\alias{format_rsl_excel}
\title{Convert complex rsl Source xlsx files into formatted rsl xlsx source files.}
\usage{
format_rsl_excel(x, thq_x, y, thq_y, z)
}
\arguments{
\item{x}{./rsl/rsl_files/rsl_thq1_nov_2021.xlsx - The downloaded xls summary table for THQ 1 from RSL-Generic Tables (2020)}

\item{thq_x}{corresponding thq for x, which is 1 here.}

\item{thq_y}{corresponding thq for y, which is 0.1 here.}

\item{Y}{./rsl/rsl_files/rsl_thq0.1_nov_2021.xlsx - The downloaded xls summary table for THQ 0.1 from RSL-Generic Tables (2020)}

\item{Z}{./rsl/rsl_files/rsl_subchronic_nov_2021.xlsx- downloaded subchronic toxicity table}
}
\description{
Convert complex rsl Source xlsx files into formatted rsl xlsx source files.
}
