% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.originals.R
\name{generate.originals}
\alias{generate.originals}
\title{Duplicate any columns with '_original'
Set Toxval Defaults}
\usage{
generate.originals(toxval.db, mat)
}
\arguments{
\item{toxval.db}{The version of toxval from which to set defaults.}

\item{mat}{THe matrix of data to be altered}
}
\value{
The altered input matrix
}
\description{
Duplicate any columns with '_original'
Set Toxval Defaults
}
