% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.chemical.info.by.source.combined.R
\name{get.chemical.info.by.source.combined}
\alias{get.chemical.info.by.source.combined}
\title{get.chemical.info.by.source.combined}
\usage{
get.chemical.info.by.source.combined(source.db, source_table, source)
}
\arguments{
\item{source.db}{The version of toxval source to use.}

\item{source_table}{The name of toxval source table to use.}

\item{source}{The name of toxval source to use.}
}
\value{
database info collected
}
\description{
get chemical info from source db tables for curation, create
chemical table to map curated chemicals to.
}
