% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.chemical.info.by.source.R
\name{get.chemical.info.by.source}
\alias{get.chemical.info.by.source}
\title{FUNCTION_TITLE}
\usage{
get.chemical.info.by.source(source.db, source_table, source, file_id)
}
\arguments{
\item{source.db}{The version of toxval source to use.}

\item{source_table}{The name of toxval source table to use.}

\item{source}{The name of toxval source to use.}

\item{file_id}{The suffixed 5 digit identifiers specified in the file names in the folder ./chemical_mapping/source_chemical_files}
}
\value{
database info collected
}
\description{
get chemical info from source db tables
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
