% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cid.list.toxval.R
\name{get.cid.list.toxval}
\alias{get.cid.list.toxval}
\title{Get chemical ids for many given CASRN/Chemical name pairs}
\usage{
get.cid.list.toxval(toxval.db, chemical.list, source, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval that the chemical id is pulled from.}

\item{chemical.list}{A 2-column dataframe of CAS Registry Numbers and chemical names.}

\item{source}{The source of the chemical data}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\value{
A 3-column dataframe of CAS Registry Numbers, chemical names, and associated chemical IDs.
}
\description{
Get chemical ids for many given CASRN/Chemical name pairs
}
