% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.clowder.file.maps.R
\name{get.clowder.file.maps}
\alias{get.clowder.file.maps}
\title{get.clowder.file.maps}
\usage{
get.clowder.file.maps(apiKey)
}
\arguments{
\item{apiKey}{User Clowder API key to access Clowder}
}
\value{
Dataframe list of Clowder ID maps
}
\description{
Pull Clowder file information to map to toxval records
}
