% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.num.decimal.count.R
\name{get.num.decimal.count}
\alias{get.num.decimal.count}
\title{get.num.decimal.count}
\usage{
get.num.decimal.count(in_num)
}
\arguments{
\item{in_num}{PARAM_DESCRIPTION}
}
\value{
Returns a dataframe of the length of the numeric and decimal places
}
\description{
A function count the length of an input numeric and its decimal places
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_count}}
}
