% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clowder_docList.R
\name{get_clowder_docList}
\alias{get_clowder_docList}
\title{Get Clowder Document List}
\usage{
get_clowder_docList(apiKey, clowder_dataset)
}
\arguments{
\item{apiKey}{The API key required for a user to access the Clowder dataset}

\item{clowder_dataset}{A character string for the dataset name}
}
\value{
Returns a dataframe with file details of: file size, date_created, file type, file id, and filename
}
\description{
This is a helper function to get a list of documents available in a Clowder dataset
}
