% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBConn.R
\name{getDBConn}
\alias{getDBConn}
\title{Get the names the database server, user, and pass or returns error message}
\usage{
getDBConn()
}
\value{
print the database connection information
}
\description{
Get the names the database server, user, and pass or returns error message
}
