% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSQLDBConn.R
\name{getPSQLDBConn}
\alias{getPSQLDBConn}
\title{Get the names the database server, user, and pass or returns error message}
\usage{
getPSQLDBConn()
}
\value{
print the database connection information
}
\description{
Get the names the database server, user, and pass or returns error message
}
