% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heast_import_source.R
\name{heast.load.all}
\alias{heast.load.all}
\title{Load HEAST to toxval_source. The data to be loaded is in the file
./heast/heast_files/EPA_HEAST_Table1_ORNL for loading.xlsx}
\usage{
heast.load.all(source.db)
}
\arguments{
\item{source.db}{The version of toxval_source into which the tables are loaded.}
}
\description{
Load HEAST to toxval_source. The data to be loaded is in the file
./heast/heast_files/EPA_HEAST_Table1_ORNL for loading.xlsx
}
