% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hess_record_url_from_clowder.R
\name{hess_record_url_from_clowder}
\alias{hess_record_url_from_clowder}
\title{Get record urls for hess documents from clowder}
\usage{
hess_record_url_from_clowder(toxval.db, source.db, apiKey, clowder_dataset)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{source.db}{The version of toxval source into which clowder data is loaded.}

\item{apiKey}{The API key required for a user to access the Clowder dataset, prefix the apiKey value with "?key="}

\item{clowder_dataset}{A character string for the dataset name
apiKey="", clowder_dataset="CCTE ToxValDB PDFs"}
}
\description{
Get record urls for hess documents from clowder
}
