% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.envirotox.source.R
\name{import_envirotox_source}
\alias{import_envirotox_source}
\title{FUNCTION_TITLE}
\usage{
import_envirotox_source(
  db,
  infile = "envirotox_taxonomy clean casrn.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{infile}{The input file ./envirotox/envirotox_files/envirotox_taxonomy.xlsx}

\item{toxval.db}{The version of toxval into which the source info is loaded.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load EnviroTox.V2 Source data into dev_toxval_source_v4.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
