% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.source.info.by.source.R
\name{import.source.info.by.source}
\alias{import.source.info.by.source}
\title{Load Source Info for each source into toxval
The information is in the file ~/dictionary/source_in_2020_aug_17.xlsx}
\usage{
import.source.info.by.source(toxval.db, source = NULL)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{source}{The specific source to be loaded, If NULL, load for all sources}
}
\description{
Load Source Info for each source into toxval
The information is in the file ~/dictionary/source_in_2020_aug_17.xlsx
}
