% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.source.info.R
\name{import.source.info}
\alias{import.source.info}
\title{Load Source Info into toxval. 
The information is in the file ./dictionary/source_in_2020_aug_17.xlsx}
\usage{
import.source.info(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}
}
\description{
Load Source Info into toxval. 
The information is in the file ./dictionary/source_in_2020_aug_17.xlsx
}
