% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atsdr_pfas_2021_source.R
\name{import_atsdr_pfas_2021_source}
\alias{import_atsdr_pfas_2021_source}
\title{Load atsdr pfas 2021 Source into dev_toxval_source_v4.}
\usage{
import_atsdr_pfas_2021_source(toxval.db, filepath)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{filepath}{The path for all the input xlsx files ./atsdr_pfas_2021/atsdr_pfas_2021_files}
}
\description{
Load atsdr pfas 2021 Source into dev_toxval_source_v4.
}
