% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atsdr_pfas_source.R
\name{import_atsdr_pfas_source}
\alias{import_atsdr_pfas_source}
\title{Load atsdr pfas Source files into dev_toxval_source_v3.}
\usage{
import_atsdr_pfas_source(
  toxval.db,
  infile1,
  infile2,
  infile3,
  infile4,
  infile5
)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile1}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_Perfluoroalkyls_Inhalation.xlsx}

\item{infile2}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_Perfluoroalkyls_Oral.xlsx}

\item{infile3}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_PFOA_Inhalation.xlsx}

\item{infile4}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_PFOA_Oral.xlsx}

\item{infile5}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_PFOS_Oral.xlsx}
}
\description{
Load atsdr pfas Source files into dev_toxval_source_v3.
}
