% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atsdr_source.R
\name{import_atsdr_source}
\alias{import_atsdr_source}
\title{Load atsdr Source into dev_toxval_source_v3.}
\usage{
import_atsdr_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./atsdr/atsdr_files/ATSDR_MRLs_2020_Sept2020_Temp.xlsx}
}
\description{
Load atsdr Source into dev_toxval_source_v3.
}
