% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_caloehha_source.R
\name{import_caloehha_source}
\alias{import_caloehha_source}
\title{Load caloehha Source file into dev_toxval_source_v4. 
The raw data can be exported as an Excel sheet from the web site
https://oehha.ca.gov/chemicals, selecting the link "Export database as .CSV file"}
\usage{
import_caloehha_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./caloehha/caloehha_files/OEHHA-chemicals_2018-10-30T08-50-47.xlsx}
}
\description{
This method parses that file and prepares for loading into toxval source
}
