% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chiu_source.R
\name{import_chiu_source}
\alias{import_chiu_source}
\title{Load chiu Source into dev_toxval_source_v3. 
Data from the Chiu et al. paper on RfD values}
\usage{
import_chiu_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./chiu/chiu_files/Full_RfD_databaseQAed-FINAL.xlsx}
}
\description{
Load chiu Source into dev_toxval_source_v3. 
Data from the Chiu et al. paper on RfD values
}
