% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_copper_source.R
\name{import_copper_source}
\alias{import_copper_source}
\title{Load copper manufacturers Source into dev_toxval_source_v4.}
\usage{
import_copper_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./copper/copper_files/Copper Data Entry - Final.xlsx}
}
\description{
Load copper manufacturers Source into dev_toxval_source_v4.
}
