% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_profile_dups_log.R
\name{import.dup.log.info}
\alias{import.dup.log.info}
\title{FUNCTION_TITLE}
\usage{
import.dup.log.info(db, dups_log_file = "True_Duplicate_for_DAT.xlsx")
}
\arguments{
\item{db}{The version of toxval into which the source info is loaded.}

\item{dups_log_file}{The name of the duplicates log file to load}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
#' Data Profiling Dups Log
Load Source Info into toxval source.
The information is in the file ./data_profile/data_profile_files/data_profiling_dups_log3.xlsx

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[read_xlsx][readxl::read_xlsx]
 [separate_rows][tidyr::separate_rows], [unite][tidyr::unite]
 [filter][dplyr::filter], [select][dplyr::select]
}
