% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dod_source.R
\name{import_dod_source}
\alias{import_dod_source}
\title{Load DOD MEG to toxval_source. The file to be laoded are in ./dod/dod_files}
\usage{
import_dod_source(source.db)
}
\arguments{
\item{source.db}{The version of toxval_source into which the tables are loaded.}
}
\description{
Load DOD MEG to toxval_source. The file to be laoded are in ./dod/dod_files
}
