% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doe_benchmarks_source.R
\name{import_doe_benchmarks_source}
\alias{import_doe_benchmarks_source}
\title{FUNCTION_TITLE}
\usage{
import_doe_benchmarks_source(
  db,
  infile = "DOE_Wildlife_Benchmarks_1996.xlsx",
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./doe_benchmarks/doe_benchmarks_files/DOE_Wildlife_Benchmarks_1996.xlsx}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load doe_benchmarks Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
