% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doe_lanl_ecorisk_source.R
\name{import_doe_lanl_ecorisk_source}
\alias{import_doe_lanl_ecorisk_source}
\title{import_doe_lanl_ecorisk_source}
\usage{
import_doe_lanl_ecorisk_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Import doe_lanl_ecorisk source into ToxVal
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_remove}}
 \code{\link[stringr]{fixed}}, \code{\link[stringr]{str_squish}}
 \code{\link[tidyr]{pivot_longer}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{rename}}
 \code{\link[tidyselect]{all_of}}
}
