% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doe_pac_source.R
\name{import_doe_pac_source}
\alias{import_doe_pac_source}
\title{FUNCTION_TITLE}
\usage{
import_doe_pac_source(db, infile = "Revision_29.xlsx", chem.check.halt = F)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./doe/doe_files/Revision_29.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load DOE Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter-joins}}
 \code{\link[tidyr]{reexports}}
}
