% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doe_source.R
\name{import_doe_source}
\alias{import_doe_source}
\title{FUNCTION_TITLE}
\usage{
import_doe_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./doe/doe_files/Revision_29.xlsx}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load doe Source into dev_toxval_source_v4.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter-joins}}
}
