% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_echa3_source.R
\name{import_echa3_source}
\alias{import_echa3_source}
\title{FUNCTION_TITLE}
\usage{
import_echa3_source(
  toxval.db,
  infile,
  dict_toxval_units,
  dict_toxval_type,
  dict_study_type,
  dict_exposure_route,
  dict_exposure_method,
  dict_study_duration,
  dict_species,
  dict_critical_effects
)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./echa3/echa3_files/TSCA_POC_Chemical_Results_081220.xlsx}

\item{dict_toxval_units}{The input file ./echa3/echa3_files/toxval_units_dictionary.xlsx}

\item{dict_toxval_type}{The input file ./echa3/echa3_files/toxval_type_dictionary.xlsx}

\item{dict_study_type}{The input file ./echa3/echa3_files/study_type_dictionary.xlsx}

\item{dict_exposure_route}{The input file ./echa3/echa3_files/exposure_route_dictionary.xlsx}

\item{dict_exposure_method}{The input file ./echa3/echa3_files/exposure_method_dictionary.xlsx}

\item{dict_study_duration}{The input file ./echa3/echa3_files/study_duration_dictionary.xlsx}

\item{dict_species}{The input file ./echa3/echa3_files/echa3_species_dict.xlsx}

\item{dict_critical_effects}{The input file ./echa3/echa3_files/critical_effect_dictionary.xlsx}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load ECHA TSCA POC Source into dev_toxval_source_v4.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[janitor]{excel_numeric_to_date}}
 \code{\link[stringr]{str_extract}}, \code{\link[stringr]{modifiers}}, \code{\link[stringr]{str_split}}
}
