% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_echa_echemportal_api_source.R
\name{import_echa_echemportal_api_source}
\alias{import_echa_echemportal_api_source}
\title{FUNCTION_TITLE}
\usage{
import_echa_echemportal_api_source(
  db,
  filepath = "echa_echemportal_api/echa_echemportal_api_files",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{filepath}{The path for all the input xlsx files ./echa_echemportal_api/echa_echemportal_api_files}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load ECHA echemportal api Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[utils]{type.convert}}
 \code{\link[data.table]{rbindlist}}
}
