% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_echa_source.R
\name{import_echa_source}
\alias{import_echa_source}
\title{Load ECHA Source from dev_toxval_source_v4(used in v8) saved as infile to dev_toxval_source_v4}
\usage{
import_echa_source(toxval.db, infile, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./echa/echa_files/echa_raw.xlsx}
}
\description{
Load ECHA Source from dev_toxval_source_v4(used in v8) saved as infile to dev_toxval_source_v4
}
