% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_efsa2_source.R
\name{import_efsa2_source}
\alias{import_efsa2_source}
\title{Load efsa2 Source into dev_toxval_source_v2.}
\usage{
import_efsa2_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./efsa2/efsa2_files/merge2/EFSA_combined_new.xlsx}
}
\description{
Load efsa2 Source into dev_toxval_source_v2.
}
