% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_efsa_source.R
\name{import_efsa_source}
\alias{import_efsa_source}
\title{Process the raw excel files downloaded from EFSA version3(March 27 2020)
To get the files, go to the web site
https://zenodo.org/record/3693783#.XrsBMmhKjIU. At the bottom are links to
a set of Excel files - download all of them into the next version V3,
convert the xlsx files to csv since the xlsx files when read in R converts all symbols/special characters
to certain unicode values (space to '_x0020_'). while reading the original xlsx files into R
it was unsuccessful to convert enconding to UTF-8, also tried converting using stringi. Only workaround was
by converting the downloaded files to csv and using the csv files as input source.}
\usage{
import_efsa_source(db, chem.check.halt = F)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Process the raw excel files downloaded from EFSA version3(March 27 2020)
To get the files, go to the web site
https://zenodo.org/record/3693783#.XrsBMmhKjIU. At the bottom are links to
a set of Excel files - download all of them into the next version V3,
convert the xlsx files to csv since the xlsx files when read in R converts all symbols/special characters
to certain unicode values (space to '_x0020_'). while reading the original xlsx files into R
it was unsuccessful to convert enconding to UTF-8, also tried converting using stringi. Only workaround was
by converting the downloaded files to csv and using the csv files as input source.
}
