% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.envirotox.source.R
\name{import_envirotox_source}
\alias{import_envirotox_source}
\title{Load EnviroTox.V2 Source data into dev_toxval_source_v4.}
\usage{
import_envirotox_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{infile}{The input file ./envirotox/envirotox_files/envirotox_taxonomy.xlsx}
}
\description{
Load EnviroTox.V2 Source data into dev_toxval_source_v4.
}
