% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_caloehha_rel_derivations_source.R,
%   R/import_echa_rac_oel_source.R, R/import_epa_dcap_source.R,
%   R/import_epa_etap_source.R, R/import_epa_hawc_source_orchestrate.R,
%   R/import_epa_hwir_source.R, R/import_epa_tsca_8e_source.R,
%   R/import_eu_scoel_source.R, R/import_generic_source.R,
%   R/import_il_epa_source.R, R/import_mass_orsg_source.R,
%   R/import_niosh_idlh_source.R, R/import_nj_dep_source.R,
%   R/import_source_usgs_hbsl.R, R/import_source_who_ipcs.R,
%   R/import_source_who_jecfa_adi.R, R/import_tx_tceq_source.R,
%   R/import_vt_vdh_dwg_source.R, R/import_who_dwg_source.R
\name{import_caloehha_rel_derivations_source}
\alias{import_caloehha_rel_derivations_source}
\alias{import_echa_rac_oel_source}
\alias{import_epa_dcap_source}
\alias{import_epa_etap_source}
\alias{import_epa_hawc_source_orchestrate}
\alias{import_epa_hwir_source}
\alias{import_epa_tsca_8e_source}
\alias{import_eu_scoel_source}
\alias{import_generic_source}
\alias{import_il_epa_source}
\alias{import_mass_orsg_source}
\alias{import_niosh_idlh_source}
\alias{import_nj_dep_source}
\alias{import_source_usgs_hbsl}
\alias{import_who_ipcs}
\alias{import_source_who_jecfa_adi}
\alias{import_tx_tceq_source}
\alias{import_vt_vdh_dwg_source}
\alias{import_who_dwg_source}
\title{import_caloehha_rel_derivations_source}
\usage{
import_caloehha_rel_derivations_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_echa_rac_oel_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_epa_dcap_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_epa_etap_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_epa_hawc_source_orchestrate(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_epa_hwir_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_epa_tsca_8e_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_eu_scoel_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_generic_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_il_epa_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_mass_orsg_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_niosh_idlh_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_nj_dep_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_source_usgs_hbsl(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_who_ipcs(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_source_who_jecfa_adi(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_tx_tceq_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_vt_vdh_dwg_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_who_dwg_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is pushed to toxval_source

None; data is pushed to toxval_source

None. SQL statements executed.

OUTPUT_DESCRIPTION

None; data is pushed to toxval_source

None; data is pushed to toxval_source

None; data is pushed to toxval_source

None; data is pushed to toxval_source

None. SQL statements are run to load data to toxval_source

None; data is pushed to toxval_source

None; data is pushed to toxval_source

None; data is pushed to toxval_source

None; data is pushed to toxval_source

OUTPUT_DESCRIPTION

None. Data is processed into the toxval_source database

OUTPUT_DESCRIPTION

None; data is pushed to toxval_source

None; data is pushed to toxval_source

None; data is pushed to toxval_source
}
\description{
Import of Cal OEHHA REL Derivations source into toxval_source

Import of ECHA EAC OEL source into toxval_source

Import of EPA DCAP 2025-05-07 source into toxval_source

Import of EPA ETAP 2024-03-07 source into toxval_source

Import of IRIS 2023-05-09 source into toxval_source

Import of EPA HWIR source into toxval_source

Import of EPA TSCA 8e source into toxval_source

Import of EU SCOEL source into toxval_source

A generic template for adding data to toxval_source for a new source

Import of IL EPA source into toxval_source

Import of Mass. ORSG source into toxval_source

Import of NIOSH IDLH source into toxval_source

Import of NJ DEP source into toxval_source

Import USGS HBSL data into toxval_source

Import of WHO IPCS data

Import of WHO JECFA ADI data

Import of TX TCEQ source into toxval_source

Import VT VDH DWG source into toxval_source

Import of WHO DWG source into toxval_source
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}
}
