% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hawc_pfas_150_source.R
\name{import_hawc_pfas_150_source}
\alias{import_hawc_pfas_150_source}
\title{Load HAWC PFAS 150 Source into toxval_source}
\usage{
import_hawc_pfas_150_source(
  db,
  infile1 = "hawc_pfas_150_raw3.xlsx",
  infile2 = "hawc_pfas_150_doses3.xlsx",
  infile3 = "hawc_pfas_150_groups3.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile1}{The input file ./hawc_pfas/hawc_pfas_files/hawc_pfas_150_raw3.xlsx , extracted
from https://hawcprd.epa.gov , assessment name - PFAS 150 (2021) and assessment id - 100500085.
Data extraction using HawcClient and extraction script hawc_pfas_150.py}

\item{infile2}{The input file ./hawc_pfas/hawc_pfas_files/hawc_pfas_150_doses3.xlsx}

\item{infile3}{The input file ./hawc_pfas/hawc_pfas_files/hawc_pfas_150_groups3.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load HAWC PFAS 150 Source into toxval_source
}
