% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hawc_pfas_430_source.R
\name{import_hawc_pfas_430_source}
\alias{import_hawc_pfas_430_source}
\title{Load HAWC PFAS 430 Source into dev_toxval_source_v4.}
\usage{
import_hawc_pfas_430_source(toxval.db, infile1, infile2, infile3)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile1}{The input file ./hawc_pfas_430/hawc_pfas_430_files/hawc_pfas_430_raw3.xlsx , extracted
from https://hawcprd.epa.gov , assessment name - PFAS 430 (2020) and assessment id - 100500256. 
Data extraction using HawcClient and extraction script hawc_pfas_430.py}

\item{infile2}{The input file ./hawc_pfas_430/hawc_pfas_430_files/hawc_pfas_430_doses3.xlsx}

\item{infile3}{The input file ./hawc_pfas_430/hawc_pfas_430_files/hawc_pfas_430_groups3.xlsx}
}
\description{
Load HAWC PFAS 430 Source into dev_toxval_source_v4.
}
