% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hawc_pfas_source.R
\name{import_hawc_pfas_source}
\alias{import_hawc_pfas_source}
\title{FUNCTION_TITLE}
\usage{
import_hawc_pfas_source(db, hawc_num = NULL, chem.check.halt = F)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{hawc_num}{The HAWC number being processed (e.g. 150, 430)}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load HAWC PFAS Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[dplyr]{select}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}
 \code{\link[tidyr]{pivot_wider}}, \code{\link[tidyr]{unite}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}
 \code{\link[stringr]{str_trim}}
 \code{\link[purrr]{map2}}
 \code{\link[digest]{digest}}
}
