% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hawc_source.R
\name{import_hawc_source}
\alias{import_hawc_source}
\title{FUNCTION_TITLE}
\usage{
import_hawc_source(
  db,
  infile1 = "hawc_original_12_06_21.xlsx",
  infile2 = "dose_dict.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile1}{The input file ./hawc/hawc_files/hawc_original_12_06_21.xlsx}

\item{infile2}{The input file ./hawc/hawc_files/dose_dict.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load HAWC Source into toxval_source

Note that the different tabs in the input sheet have different names, so these need
to be adjusted manually for the code to work. This is a problem wit how the data
is stored in HAWC
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{getSheetNames}}, \code{\link[openxlsx]{read.xlsx}}
 \code{\link[stats]{aggregate}}
 \code{\link[digest]{digest}}
}
