% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_health_canada_source.R
\name{import_health_canada_source}
\alias{import_health_canada_source}
\title{Load Health Canada Source Info into toxval_source}
\usage{
import_health_canada_source(
  db,
  infile = "HealthCanada_TRVs_2010_AppendixA v2.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{infile}{The input file ./health_canada/health_canada_files/HealthCanada_TRVs_2010_AppendixA v2.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load Health Canada Source Info into toxval_source
}
