% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_heast_source.R
\name{import_heast_source}
\alias{import_heast_source}
\title{FUNCTION_TITLE}
\usage{
import_heast_source(
  db,
  infile = "EPA_HEAST_Table1_ORNL for loading.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./heast/heast_files/EPA_HEAST_Table1_ORNL for loading.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load HEAST Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
