% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hess_source.R
\name{import_hess_source}
\alias{import_hess_source}
\title{Load hess Source into dev_toxval_source_v3.}
\usage{
import_hess_source(toxval.db, infile1, infile2)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile1}{The input file ./hess/hess_files/hess_6_16_21.csv, extracted by Risa Sayre(SCDCD)}

\item{infile2}{The input file ./hess/hess_files/hess_record_urls_from_clowder.xlsx}
}
\description{
Load hess Source into dev_toxval_source_v3.
}
