% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hpvis_source.R
\name{import_hpvis_source}
\alias{import_hpvis_source}
\title{Load hpvis Source Info into dev_toxval_source_v2.}
\usage{
import_hpvis_source(toxval.db, filepath)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{filepath}{The path for all the input xlsx files ./hpvis/hpvis_files}
}
\description{
Load hpvis Source Info into dev_toxval_source_v2.
}
