% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_iris_source.R
\name{import_iris_source}
\alias{import_iris_source}
\title{Load IRIS Source into dev_toxval_source_v4.}
\usage{
import_iris_source(toxval.db, infile1, infile2)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile1}{The input file ./iris/iris_files/IRIS_non_cancer_clean 2020-05-27.xlsx}

\item{infile2}{The input file ./iris/iris_files/IRIS_cancer_clean 2020-05-27.xlsx}
}
\description{
Load IRIS Source into dev_toxval_source_v4.
}
