% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_lanl_source.R
\name{import_lanl_source}
\alias{import_lanl_source}
\title{Load LANL Source into toxval_source}
\usage{
import_lanl_source(db, infile = "ESLs_R3.3.xlsx", chem.check.halt = T)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./lanl/lanl_files/ESLs_R3.3.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load LANL Source into toxval_source
}
