% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_niosh_source.R
\name{import_niosh_source}
\alias{import_niosh_source}
\title{Load NIOSH Source into toxval_source}
\usage{
import_niosh_source(db, infile = "niosh_IDLH_2020.xlsx", chem.check.halt = T)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./niosh/niosh_files/niosh_IDLH_2020.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load NIOSH Source into toxval_source
}
