% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_opp_source.R
\name{import_opp_source}
\alias{import_opp_source}
\title{import_opp_source}
\usage{
import_opp_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is pushed to toxval_source
}
\description{
Load EPA OPP data to toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}
 \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{row_number}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{bind_rows}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_extract}}
}
