% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_penn_source.R
\name{import_penn_source}
\alias{import_penn_source}
\title{Load Penn Source into toxval_source}
\usage{
import_penn_source(db, infile = "..enn DEP Table 5a.xlsx", chem.check.halt = T)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./penn/penn_files/Penn DEP Table 5a.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load Penn Source into toxval_source
}
