% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pfas_150_sem_source.R
\name{import_pfas_150_sem_source}
\alias{import_pfas_150_sem_source}
\title{Load PFAS 150 SEM Source data into toxval_source}
\usage{
import_pfas_150_sem_source(
  db,
 
    infile = "PFAS150 animal study template combined_clearance with DTXSID and CASRN.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{infile}{The input file ./PFAS 150 SEM/PFAS 150 SEM_files/PFAS150 animal study template combined_clearance with DTXSID and CASRN.xlsx}

\item{chem.check.halt}{If TRUE and there are problems with chemicals CASRN checks, halt the program}
}
\description{
Load PFAS 150 SEM Source data into toxval_source
}
