% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pfas_summary_pods_source.R
\name{import_pfas_summary_pods_source}
\alias{import_pfas_summary_pods_source}
\title{Load PFAS Summary PODs into dev_toxval_source_v2.}
\usage{
import_pfas_summary_pods_source(toxval.db, infile1, infile2)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile1}{The input file ./PFAS Summary PODs/PFAS Summary PODs_files/PFAS 150 Study Level PODs_061920.xlsx}

\item{infile2}{The input file ./PFAS Summary PODs/PFAS Summary PODs_files/CompToxChemicalsDashboard-Batch-Search_2020-07-20_17_18_42.xls}
}
\description{
Load PFAS Summary PODs into dev_toxval_source_v2.
}
