% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pprtv_ncea_source.R
\name{import_pprtv_ncea_source}
\alias{import_pprtv_ncea_source}
\title{Load pprtv_ncea Source Info into dev_toxval_source_v2.}
\usage{
import_pprtv_ncea_source(toxval.db, filepath, csvfile, scrapepath)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{filepath}{The path for all the input xlsx files ./pprtv_ncea/pprtv_ncea_files}

\item{csvfile}{The input csv file ./pprtv_ncea/pprtv_ncea_files/dose_reg2.csv}

\item{scrapepath}{The path for new_pprtv_ncea_scrape_table file ./pprtv_ncea/PPRTV_scrape2020-04-08.xlsx}
}
\description{
Load pprtv_ncea Source Info into dev_toxval_source_v2.
}
