% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rsl_source_info.R
\name{import_rsl_source_info}
\alias{import_rsl_source_info}
\title{Load rsl Source Info into dev_toxval_source_v2.}
\usage{
import_rsl_source_info(toxval.db, infile1a, infile1b, infile2, infile3)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{infile1a}{The input file ./rsl/rsl_files/final_rsl_thq_combined_nov21.xlsx}

\item{infile1b}{The input file ./rsl/rsl_files/final_rsl_subchronic_nov21.xlsx}

\item{infile2}{The input file ./rsl/rsl_files/general_info_nov_21.xlsx}

\item{infile3}{The input file ./rsl/rsl_files/key_description_nov_21.xlsx}
}
\description{
Load rsl Source Info into dev_toxval_source_v2.
}
