% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_atsdr_mrls.R
\name{import_source_atsdr_mrls}
\alias{import_source_atsdr_mrls}
\title{import_source_astdr_mrls}
\usage{
import_source_atsdr_mrls(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE,
  do.toxicological_profile = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}

\item{do.toxicological_profile}{If TRUE, add toxicological profile data to table before insertion}
}
\value{
None; data is pushed to toxval_source
}
\description{
Send ASTDR MRLs data to toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[tidyr]{separate}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}
 \code{\link[stringr]{str_trim}}
}
