% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_atsdr_mrls_2022.R
\name{import_generic_source}
\alias{import_generic_source}
\title{FUNCTION_TITLE}
\usage{
import_generic_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Import of ATSDR MRLs 2022 source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[stringr]{str_trim}}
 \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{recode}}
 \code{\link[tidyr]{pivot_longer}}
}
