% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cosmos_source.R
\name{import_source_cosmos}
\alias{import_source_cosmos}
\title{import_source_cosmos}
\usage{
import_source_cosmos(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}

\item{do.reset}{If TRUE, delete data from the database for this source before inserting new data}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is loaded into toxval_source
}
\description{
Load COSMOS source data into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{filter}}
 \code{\link[stringr]{str_trim}}
 \code{\link[tidyr]{separate_rows}}, \code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{separate}}
}
