% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_epa_aegl.R
\name{import_source_epa_aegl}
\alias{import_source_epa_aegl}
\title{import_source_epa_aegl}
\usage{
import_source_epa_aegl(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{PARAM_DESCRIPTION}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Import EPA AEGL data into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{mutate-joins}}
 \code{\link[tidyr]{separate_rows}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{reexports}}
 \code{\link[stringr]{modifiers}}, \code{\link[stringr]{str_remove}}, \code{\link[stringr]{str_split}}, \code{\link[stringr]{str_trim}}
}
