% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_epa_hhtv.R
\name{import_source_epa_hhtv}
\alias{import_source_epa_hhtv}
\title{import_source_epa_hhtv}
\usage{
import_source_epa_hhtv(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is pushed to toxval_source
}
\description{
Push EPA HHTV data to toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}
}
