% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_epa_ow_npdwr.R
\name{import_source_epa_ow_npdwr}
\alias{import_source_epa_ow_npdwr}
\title{FUNCTION_TITLE}
\usage{
import_source_epa_ow_npdwr(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
#' Import of EPA OW NPDWR source into toxval_source

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}
 \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}
 \code{\link[stringr]{str_trim}}
}
