% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_epa_ow_nrwqc_alc.R
\name{import_source_epa_ow_nrwqc_alc}
\alias{import_source_epa_ow_nrwqc_alc}
\title{FUNCTION_TITLE}
\usage{
import_source_epa_ow_nrwqc_alc(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Import of EPA OW NRWQC-ALC source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}
 \code{\link[dplyr]{rename}}, \code{\link[dplyr]{c("rowwise", "rowwise", "rowwise")}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{case_when}}
 \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{reexports}}, \code{\link[tidyr]{separate_rows}}
}
