% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_iuclid_orchestrate.R
\name{import_source_iuclid_orchestrate}
\alias{import_source_iuclid_orchestrate}
\title{import_source_iuclid_orchestrate}
\usage{
import_source_iuclid_orchestrate(
  dir = paste0(toxval.config()$datapath, "iuclid")
)
}
\arguments{
\item{dir}{directory containing the various IUCLID subsource subdirectories}

\item{db}{The version of toxval_source into which the source is loaded.}

\item{do.insert}{If TRUE, insert data into the database, default TRUE}

\item{chem.check.halt}{If TRUE, stop the execution if there are errors in the}
}
\value{
None, subsources loaded
}
\description{
Load the various IUCLID subsources into ToxVal
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
