% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_mass_mmcl.R
\name{import_source_mass_mmcl}
\alias{import_source_mass_mmcl}
\title{import_source_mass_mmcl}
\usage{
import_source_mass_mmcl(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}

\item{do.reset}{If TRUE, delete data from the database for this source before inserting new data}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is loaded into toxval_source
}
\description{
Load Mass. Drinking Water Standards into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{coalesce}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{case_when}}
 \code{\link[purrr]{reexports}}
 \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{separate}}
 \code{\link[stringr]{str_match}}, \code{\link[stringr]{str_trim}}
}
