% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_penn_dep_mcls.R
\name{import_source_penn_dep_mcls}
\alias{import_source_penn_dep_mcls}
\title{import_source_penn_dep}
\usage{
import_source_penn_dep_mcls(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is loaded to MySQL server
}
\description{
Load Pennsylvania DEP MCLs into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{distinct}}
 \code{\link[stringr]{str_trim}}
 \code{\link[tidyr]{unite}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}
}
