% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_who_jecfa_tox_studies.R
\name{import_source_who_jecfa_tox_studies}
\alias{import_source_who_jecfa_tox_studies}
\title{import_source_who_jecfa_tox_studies}
\usage{
import_source_who_jecfa_tox_studies(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALS}
}
\value{
None; data is pushed to toxval_source
}
\description{
Import of WHO JECFA Tox Studies data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{filter}}
 \code{\link[tidyr]{separate_rows}}
 \code{\link[stringr]{str_trim}}
}
