% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_test_source.R
\name{import_test_source}
\alias{import_test_source}
\title{import_test_source}
\usage{
import_test_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}

\item{do.reset}{If TRUE, delete data from the database for this source before inserting new data}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is loaded into toxval_source
}
\description{
Load TEST Source data into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[utils]{read.table}}
 \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{join_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}
 \code{\link[stringr]{str_trim}}
 \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{drop_na}}
}
