% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_wignall_source.R
\name{import_wignall_source}
\alias{import_wignall_source}
\title{Load wignall Source data into dev_toxval_source_v2.}
\usage{
import_wignall_source(
  db,
  infile = "BMD_Results_2014-06-17_reviewed Mar 2018 parsed.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval into which the source info is loaded.}

\item{infile}{The input file ./wignall/wignall_files/BMD_Results_2014-06-17_reviewed Mar 2018.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load wignall Source data into dev_toxval_source_v2.
}
