% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.audit.table.R
\name{init.audit.table}
\alias{init.audit.table}
\title{Create audit table and add BEFORE UPDATE audit triggers to source_* tables}
\usage{
init.audit.table(db, do.halt = FALSE, verbose = FALSE)
}
\arguments{
\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}
}
\description{
Create audit table and add BEFORE UPDATE audit triggers to source_* tables
}
