% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_new_source_dir.R
\name{init_new_source_dir}
\alias{init_new_source_dir}
\title{Initialize New Source Directory}
\usage{
init_new_source_dir(source_table = NULL)
}
\arguments{
\item{source_table}{Character string or vector of source tables names, Default: NULL}
}
\value{
None. Folders are created.
}
\description{
Create file directory for input source table names.
}
