% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_source_iuclid_directory.R
\name{initialize_source_iuclid_directory}
\alias{initialize_source_iuclid_directory}
\title{Initialize Source IUCLID Directory into subdirectory based on input files}
\usage{
initialize_source_iuclid_directory()
}
\value{
None, file directory structure generated
}
\description{
Initialize Source IUCLID Directory into subdirectory based on input files
}
