% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS_noncancer_clean.R
\name{iris.noncancer.clean}
\alias{iris.noncancer.clean}
\title{create IRIS non cancer clean source file from iris_scrape_rfd_rfc 2020-05-27.xlsx 
and iris_scrape_woe 2020-05-27.xlsx (build using iris.scraper.R)}
\usage{
iris.noncancer.clean(infile1, infile2)
}
\arguments{
\item{infile1}{The input file ../iris/iris_files/iris_scrape_rfd_rfc 2020-05-27.xlsx}

\item{infile2}{The input file ../iris/iris_files/iris_scrape_woe 2020-05-27.xlsx}
}
\description{
create IRIS non cancer clean source file from iris_scrape_rfd_rfc 2020-05-27.xlsx 
and iris_scrape_woe 2020-05-27.xlsx (build using iris.scraper.R)
}
