% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.dsstox.R
\name{load.dsstox}
\alias{load.dsstox}
\title{Load DSSTox if needed from a file into a global variables (DSSTOX)}
\usage{
load.dsstox()
}
\description{
Load DSSTox if needed from a file into a global variables (DSSTOX)
}
