% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clowder_document_name.R
\name{load_clowder_document_name}
\alias{load_clowder_document_name}
\title{load_clowder_document_name}
\usage{
load_clowder_document_name(toxval.db, apiKey, clowder_dataset)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{apiKey}{The API key required for a user to access the Clowder dataset, prefix the apiKey value with "?key="}

\item{clowder_dataset}{A character string for the dataset name
First create a table in source db with clowder document name and record source hash
Update missing document names in toxval v9 record source table using data from clowder table
apiKey="", clowder_dataset="CCTE ToxValDB PDFs"}

\item{source.db}{The version of toxval source into which clowder data is loaded.}
}
\description{
Function to load Clowder document information to database.
}
