% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_log.R
\name{log_message}
\alias{log_message}
\title{Function to combine output log with output message}
\usage{
log_message(log_df, message_df_col)
}
\description{
Function to combine output log with output message
}
