% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript.table.sources.R
\name{manuscript.table.sources}
\alias{manuscript.table.sources}
\title{Export the table of the sources for the manuscript
#' @param toxval.db Database version}
\usage{
manuscript.table.sources(toxval.db)
}
\description{
Export the table of the sources for the manuscript
#' @param toxval.db Database version
}
