% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.chemical.to.dsstox.by.source.R
\name{map.chemical.to.dsstox.by.source}
\alias{map.chemical.to.dsstox.by.source}
\title{Perform the DSSTox mapping}
\usage{
map.chemical.to.dsstox.by.source(toxval.db, source, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Perform the DSSTox mapping
}
