% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.chemical.to.dsstox.ecotox.R
\name{map.chemical.to.dsstox.ecotox}
\alias{map.chemical.to.dsstox.ecotox}
\title{use name and dtxsid pulled from ECOTOX in source_chemical for ecotox}
\usage{
map.chemical.to.dsstox.ecotox(toxval.db, source, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
use name and dtxsid pulled from ECOTOX in source_chemical for ecotox
}
