% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.chemical_list.to.dsstox.R
\name{map.chemical_list.to.dsstox}
\alias{map.chemical_list.to.dsstox}
\title{Perform the DSSTox mapping}
\usage{
map.chemical_list.to.dsstox(toxval.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Perform the DSSTox mapping
}
