% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.hash.record.clowder.document_name.R
\name{map.hash.record.clowder.document_name}
\alias{map.hash.record.clowder.document_name}
\title{Replace record_source_hash with clowder record_source_hash by mapping using document names}
\usage{
map.hash.record.clowder.document_name(toxval.db, apiKey, clowder_dataset)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{apiKey}{The API key required for a user to access the Clowder dataset, prefix the apiKey value with "?key="}

\item{clowder_dataset}{A character string for the dataset name
First create a table in source db with clowder document name and record source hash
Update record_source_hash in toxval v9 record source table using data from clowder table 
apiKey="?key=d2547ff7-83ee-4f4e-b7b8-9875a5b18a83", clowder_dataset="CCTE ToxValDB PDFs"}

\item{source.db}{The version of toxval source into which clowder data is loaded.}
}
\description{
Replace record_source_hash with clowder record_source_hash by mapping using document names
}
