% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.hash.record_source.by.source.R
\name{map.hash.record_source.by.source}
\alias{map.hash.record_source.by.source}
\title{Map the icf master sheet hash to record source table hash}
\usage{
map.hash.record_source.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Map the icf master sheet hash to record source table hash
}
