% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niosh_pdf_to_excel.R
\name{niosh.pdf.to.excel}
\alias{niosh.pdf.to.excel}
\title{niosh.pdf.to.excel}
\usage{
niosh.pdf.to.excel(infile)
}
\arguments{
\item{infile}{The input file ../niosh/niosh_files/CDC - Index of Chemicals - NIOSH Publications and Products.pdf}
}
\description{
Conversion of NIOSH pdf to Excel file
}
