% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_iuclid.R
\name{orchestrate_import_source_iuclid}
\alias{orchestrate_import_source_iuclid}
\title{Load the various IUCLID subsources into ToxVal}
\usage{
orchestrate_import_source_iuclid(dir = "Repo/iuclid")
}
\arguments{
\item{dir}{directory containing the various IUCLID subsource subdirectories}

\item{db}{The version of toxval_source into which the source is loaded.}

\item{do.insert}{If TRUE, insert data into the database, default TRUE}

\item{chem.check.halt}{If TRUE, stop the execution if there are errors in the}
}
\value{
None, subsources loaded
}
\description{
Load the various IUCLID subsources into ToxVal
}
