% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echemportal.prep.v2.R
\name{parse.effect.level}
\alias{parse.effect.level}
\title{Process the the string that has the qualifier, value and units}
\usage{
parse.effect.level(val)
}
\arguments{
\item{val}{The string}
}
\value{
a list with three values: qualifier, value, units
}
\description{
Process the the string that has the qualifier, value and units
}
