% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfas.by.source.R
\name{pfas.by.source}
\alias{pfas.by.source}
\title{Get the sources with PFAS data}
\usage{
pfas.by.source(db)
}
\arguments{
\item{db}{The version of toxval into which the source is loaded.}
}
\description{
Get the sources with PFAS data
}
