% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poc.summary.plot.R
\name{poc.summary.plot}
\alias{poc.summary.plot}
\title{Make a scatter plot of the relationship between info availability and Scientific Domain Metric}
\usage{
poc.summary.plot(to.file = F, sys.date, do.read = F)
}
\arguments{
\item{to.file}{If TRUE, print the graph to a file, else write to the screen}

\item{sys.date}{The date of the bin data Excel fileto read}

\item{do.read}{if TRUE, read in the source file and store in a global}
}
\description{
Make a scatter plot of the relationship between info availability and Scientific Domain Metric
}
