% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprtv.ncea.load.all.R
\name{pprtv.ncea.load.all}
\alias{pprtv.ncea.load.all}
\title{Load NEW PPRTV (NCEA) to toxval_source. The data is found in a series of files:
../pprtv_ncea/pprtv_ncea_files/assessments.xlsx
../pprtv_ncea/pprtv_ncea_files/assessment_study.xlsx
../pprtv_ncea/pprtv_ncea_files/reference.xlsx
../pprtv_ncea/pprtv_ncea_files/study.xlsx
../pprtv_ncea/pprtv_ncea_files/dose_reg2.csv
../pprtv_ncea/PPRTV_scrape2020-04-08.xlsx}
\usage{
pprtv.ncea.load.all(source.db)
}
\arguments{
\item{source.db}{The version of toxval_source into which the tables are loaded.}
}
\description{
Load NEW PPRTV (NCEA) to toxval_source. The data is found in a series of files:
../pprtv_ncea/pprtv_ncea_files/assessments.xlsx
../pprtv_ncea/pprtv_ncea_files/assessment_study.xlsx
../pprtv_ncea/pprtv_ncea_files/reference.xlsx
../pprtv_ncea/pprtv_ncea_files/study.xlsx
../pprtv_ncea/pprtv_ncea_files/dose_reg2.csv
../pprtv_ncea/PPRTV_scrape2020-04-08.xlsx
}
