% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprtv.ncea.scraper.R
\name{pprtv.ncea.scraper}
\alias{pprtv.ncea.scraper}
\title{Prepare the PPRTV data downloaded October 2018. Go to the web site
https://hhpprtv.ornl.gov/quickview/pprtv.php}
\usage{
pprtv.ncea.scraper()
}
\description{
Prepare the PPRTV data downloaded October 2018. Go to the web site
https://hhpprtv.ornl.gov/quickview/pprtv.php
}
