% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprtv.ornl.load.all.new.R
\name{pprtv.ornl.load.all.new}
\alias{pprtv.ornl.load.all.new}
\title{Prepare all of the PPRTV data from files downloaded May 2020. Go to the web site
https://hhpprtv.ornl.gov/quickview/pprtv_compare.php and download the individual
files listed here. The combined data is in the file
../pprtv_ornl/pprtv_ornl_files/PPRTV_ORNL_noncancer.xlsx}
\usage{
pprtv.ornl.load.all.new(filepath, verbose = F)
}
\arguments{
\item{filepath}{The input files are in path ../pprtv_ornl/pprtv_ornl_files}
}
\description{
Prepare all of the PPRTV data from files downloaded May 2020. Go to the web site
https://hhpprtv.ornl.gov/quickview/pprtv_compare.php and download the individual
files listed here. The combined data is in the file
../pprtv_ornl/pprtv_ornl_files/PPRTV_ORNL_noncancer.xlsx
}
