% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprtv.ornl.scraper.R
\name{pprtv.ornl.scraper}
\alias{pprtv.ornl.scraper}
\title{Prepare the PPRTV data downloaded October 2018. Go to the web site
https://hhpprtv.ornl.gov/quickview/pprtv.php}
\usage{
pprtv.ornl.scraper()
}
\description{
Prepare the PPRTV data downloaded October 2018. Go to the web site
https://hhpprtv.ornl.gov/quickview/pprtv.php
}
