% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.DAT.conversion.R
\name{prep.DAT.conversion}
\alias{prep.DAT.conversion}
\title{prep.DAT.conversion}
\usage{
prep.DAT.conversion(in_dat, hash_id_list, source)
}
\arguments{
\item{in_dat}{PARAM_DESCRIPTION}

\item{hash_id_list}{PARAM_DESCRIPTION}

\item{source}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Select and rename DAT audit columns for toxval_source, calculate new source_hash
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}
 \code{\link[tidyr]{unite}}, \code{\link[tidyr]{reexports}}
 \code{\link[purrr]{map}}
}
