% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_prescreening_summary.R
\name{qc_prescreening_summary}
\alias{qc_prescreening_summary}
\title{FUNCTION_TITLE}
\usage{
qc_prescreening_summary(src_tbl = NULL, outputDir = NULL, db = NULL)
}
\arguments{
\item{src_tbl}{a toxval source table name}

\item{outputDir}{optional directory path to save output file in}

\item{db}{the name of the database}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Runs a database query and returns a result set
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{pivot_longer}}
 \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}
 \code{\link[writexl]{write_xlsx}}
}
