% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.numeric.R
\name{reset.numeric}
\alias{reset.numeric}
\title{Reset the numeric and units values to their original values}
\usage{
reset.numeric(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}
}
\value{
The database will be altered
}
\description{
Reset the numeric and units values to their original values
}
