% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInsertTable.R
\name{runInsertTable}
\alias{runInsertTable}
\title{FUNCTION_TITLE}
\usage{
runInsertTable(mat, table, db, do.halt = T, verbose = F, get.id = T)
}
\arguments{
\item{mat}{data frame containing the data, with the column names corresponding}

\item{table}{name of the database table to which data will be inserted}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}

\item{get.id}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Inserts multiple rows into a database table
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[RMySQL]{character(0)}}, \code{\link[RMySQL]{MySQLDriver-class}}
}
