% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQueryPSQL.R
\name{runQuery_psql}
\alias{runQuery_psql}
\title{Runs a PSQL database query and returns a result set}
\usage{
runQuery_psql(query, db, do.halt = T, verbose = T)
}
\arguments{
\item{query}{a properly formatted SQL query as a string}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}
}
\description{
Runs a PSQL database query and returns a result set
}
