% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStatement.R
\name{runStatement}
\alias{runStatement}
\title{runStatement}
\usage{
runStatement(query, db, do.halt = F, verbose = F)
}
\arguments{
\item{query}{a properly formatted SQL query as a string}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}
}
\value{
None. SQL statement is run.
}
\description{
Run a SQL statement, such as an ALTER or UPDATE
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[RMySQL]{character(0)}}, \code{\link[RMySQL]{MySQLDriver-class}}
}
