% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUpdate.R
\name{runUpdate}
\alias{runUpdate}
\title{Runs a database query and returns a result set}
\usage{
runUpdate(
  table,
  updateQuery = NULL,
  updated_df = NULL,
  db,
  do.halt = TRUE,
  verbose = FALSE,
  trigger_check = TRUE
)
}
\arguments{
\item{updateQuery}{a properly formatted SQL query as a string in the form of an UPDATE INNER JOIN}

\item{updated_df}{a dataframe of updated data to temporarily write to database for INNER JOIN}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}

\item{trigger_check}{if FALSE, audit triggers are ignored/bypassed}
}
\description{
Runs a database query and returns a result set
}
