% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.dtxsid.R
\name{set.dtxsid}
\alias{set.dtxsid}
\title{Set the dtxsid values for all of the other tables}
\usage{
set.dtxsid(toxval.db, table, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{table}{the Database table to be updated}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Set the dtxsid values for all of the other tables
}
