% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.hash.bcfbaf.by.source.R
\name{set.hash.bcfbaf.by.source}
\alias{set.hash.bcfbaf.by.source}
\title{Set the hash in table bcfbaf}
\usage{
set.hash.bcfbaf.by.source(toxval.db, source, do.reset = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{do.reset}{if TRUE, reset all hashes and start rom scratch.
Otherwise only update empty values}
}
\value{
The database will be altered
}
\description{
Set the hash in table bcfbaf
}
