% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.hash.genetox_details.R
\name{set.hash.genetox_details}
\alias{set.hash.genetox_details}
\title{Set the hash in table genetox_details}
\usage{
set.hash.genetox_details(toxval.db, do.reset = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{do.reset}{if TRUE, reset all hashes and start rom scratch.
Otherwise only update empty values}
}
\value{
The database will be altered
}
\description{
Set the hash in table genetox_details
}
