% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.hash.record_source.R
\name{set.hash.record_source}
\alias{set.hash.record_source}
\title{Set the hash in table record_source}
\usage{
set.hash.record_source(toxval.db, do.reset = T)
}
\arguments{
\item{toxval.db}{The version of toxval in which the data is altered.}

\item{do.reset}{if TRUE, reset all hashes and start from scratch.
Otherwise only update empty values}
}
\value{
The database will be altered
}
\description{
Set the hash in table record_source
}
