% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.hash.source.R
\name{set.hash.source}
\alias{set.hash.source}
\title{Set the hash in source table}
\usage{
set.hash.source(source.db, source, reset = F)
}
\arguments{
\item{source.db}{The source database version where data is getting altered (eg: dev_toxval_source_v4).}

\item{source}{The table in source where the hash data is getting altered.}

\item{do.reset}{if TRUE, reset all hashes and start from scratch.
Otherwise only update empty values}
}
\value{
The database will be altered
}
\description{
Set the hash in source table
}
