% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_clowder_doc_type.R
\name{set_clowder_doc_type}
\alias{set_clowder_doc_type}
\title{set_clowder_doc_type}
\usage{
set_clowder_doc_type(
  source_table = NULL,
  source_version_date = NULL,
  clowder_url = NULL,
  clowder_api_key = NULL,
  source.db = NULL,
  ds_id = NULL,
  clowder_id_list = NULL
)
}
\arguments{
\item{source_table}{The source table name (e.g. source_test). Default is NULL for "all"}

\item{source_version_date}{The version date for the source table. Default is NULL for "all"}

\item{clowder_url}{URL to Clowder}

\item{clowder_api_key}{API key to access Clowder resources}

\item{source.db}{Name of the toxval_source database to apply updates to}

\item{ds_id}{Clowder Dataset ID for ToxVal Clowder Documents.}

\item{clowder_id_list}{Optional DataFrame with field "clowder_id" values for document records to update.}
}
\value{
None. SQL statements are performed.
}
\description{
Update documents table entries "document_type" field based on Clowder organization
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{rename}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{mutate}}
 \code{\link[tidyr]{separate_rows}}, \code{\link[tidyr]{unite}}
}
