% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_clowder_id.R
\name{set_clowder_id}
\alias{set_clowder_id}
\title{Set the clowder_id and document_name in res}
\usage{
set_clowder_id(res, source)
}
\arguments{
\item{res}{The input dataframe}

\item{source}{The data source name}
}
\value{
Returns the input dataframe with defaults set
}
\description{
Set the clowder_id and document_name in res
}
