% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_clowder_id_lineage.R
\name{set_clowder_id_lineage}
\alias{set_clowder_id_lineage}
\title{FUNCTION_TITLE}
\usage{
set_clowder_id_lineage(
  source_table,
  map_clowder_id_field,
  map_file,
  clowder_url,
  clowder_api_key,
  sync_clowder_metadata = FALSE
)
}
\arguments{
\item{source_table}{The source table name (e.g. source_test)}

\item{map_clowder_id_field}{Column name for the Clowder ID field of the map}

\item{map_file}{A dataframe of Clowder document mapping info. If NULL, will try to load a hardcoded map for the source}

\item{clowder_url}{URL to Clowder}

\item{clowder_api_key}{API key to access Clowder resources}

\item{sync_clowder_metadata}{Boolean whether to sync Clowder metadata for new document records. Default is False.}
}
\value{
Returns an updated map with newly associated toxval_source table ID values
}
\description{
Create document records and associations in toxval_source based on input source table and document map
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{rename}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{mutate}}
 \code{\link[tidyr]{separate_rows}}, \code{\link[tidyr]{unite}}
}
