% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDBConn.R
\name{setDBConn}
\alias{setDBConn}
\title{setDBConn}
\usage{
setDBConn(
  server = "ccte-mysql-res.epa.gov",
  user = NULL,
  password = NULL,
  api_auth = NULL,
  port = 3306
)
}
\arguments{
\item{server}{SQL server on which relevant database lives}

\item{user}{SQL username to access database}

\item{password}{SQL password corresponding to username}

\item{api_auth}{API Key for CCTE API's}

\item{port}{Port for the server connection. Default of 3306.}
}
\value{
None
}
\description{
set SQL connection to the database
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
