% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.size.R
\name{source.size}
\alias{source.size}
\title{print out the size of each of the tables in toxval_source}
\usage{
source.size(db = "res_toxval_source_v5")
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}
}
\description{
print out the size of each of the tables in toxval_source
}
