% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_chemical.duplicates.R
\name{source_chemical.duplicates}
\alias{source_chemical.duplicates}
\title{Find duplicated chemicals in the source_chemical table. THis will help get rid of
records that have been repalced}
\usage{
source_chemical.duplicates(db)
}
\arguments{
\item{db}{The version of toxval into which the tables are loaded.}
}
\description{
Find duplicated chemicals in the source_chemical table. THis will help get rid of
records that have been repalced
}
