% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_chemical.ecotox.R
\name{source_chemical.ecotox}
\alias{source_chemical.ecotox}
\title{FUNCTION_TITLE}
\usage{
source_chemical.ecotox(
  toxval.db,
  source.db,
  res,
  source,
  chem.check.halt = FALSE,
  casrn.col = "casrn",
  name.col = "name",
  verbose = F
)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{source.db}{The source database version}

\item{res}{PARAM_DESCRIPTION}

\item{source}{The xource to be processed (ECOTOX)}

\item{chem.check.halt}{If TRUE, halt if there are errors in the chemical checking}

\item{casrn.col}{Name of the column containing the CASRN}

\item{name.col}{Name of the column containing chemical names}

\item{verbose}{If TRUE, output extra diagnostics information}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
special process to deal with source chemicals for ECOTOX
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[digest]{digest}}
}
