% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_chemical.toxrefdb.R
\name{source_chemical.toxrefdb}
\alias{source_chemical.toxrefdb}
\title{FUNCTION_TITLE}
\usage{
source_chemical.toxrefdb(
  toxval.db,
  source.db,
  res,
  source = "ToxRefDB",
  chem.check.halt = FALSE,
  casrn.col = "casrn",
  name.col = "name",
  verbose = F
)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{source.db}{The source database version}

\item{res}{The dataframe to which the chemical_id sill be added}

\item{source}{The name of the source}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}

\item{casrn.col}{The name of the column containing the CASRN}

\item{name.col}{The name ofhte column containing hte chemical name}

\item{verbose}{If TRUE, write out diagnostic messages}
}
\value{
Returns the input dataframe with the chemical_id added
}
\description{
Special process to deal with source chemicals for ToxRefDB. This will put the
chemicals into the source database source_chemical table
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[digest]{digest}}
}
