% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_set_defaults.R
\name{source_set_defaults}
\alias{source_set_defaults}
\title{Set default value for NAs - jsut set NA to "-" for columns of type character}
\usage{
source_set_defaults(res, source)
}
\arguments{
\item{res}{The input dataframe}

\item{source}{The data source name}
}
\value{
Returns the input dataframe with defaults set
}
\description{
Set default value for NAs - jsut set NA to "-" for columns of type character
}
