% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.mapper.R
\name{species.mapper}
\alias{species.mapper}
\title{Map the species to the ECOTOX species dictionary and export the missing species to
add to the dictionary}
\usage{
species.mapper(toxval.db, date_string = "2022-02-23")
}
\arguments{
\item{toxval.db}{The version of the database to use}

\item{date_string}{The date of the dictionary versions}
}
\description{
This function replaces fix.species
This function precedes toxvaldb.load.species
}
