% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.check.source_chemical.R
\name{toxval.check.source_chemical}
\alias{toxval.check.source_chemical}
\title{Check the status of the soruce_chemical tables}
\usage{
toxval.check.source_chemical(toxval.db, source.db)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{source.db}{The vsource database version}
}
\description{
Check the status of the soruce_chemical tables
}
