% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.config.R
\name{toxval.config}
\alias{toxval.config}
\title{Define a set of global variables. These include the source path (datapath)
and the source databases (e.g. dev_toxval_{version} and dev_toxval_source_{version})
and the urls for the ACToR web services.}
\usage{
toxval.config()
}
\value{
Returns a set of parameters to be used throughout the package
}
\description{
Define a set of global variables. These include the source path (datapath)
and the source databases (e.g. dev_toxval_{version} and dev_toxval_source_{version})
and the urls for the ACToR web services.
}
