% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.init.db.R
\name{toxval.init.db}
\alias{toxval.init.db}
\title{Initialize the database. THis sill load the species, info and dictionary tables}
\usage{
toxval.init.db(toxval.db, reset = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{reset}{If TRUE, delete all content from the database}
}
\description{
Initialize the database. THis sill load the species, info and dictionary tables
}
