% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.alaska_dec.R
\name{toxval.load.alaska_dec}
\alias{toxval.load.alaska_dec}
\title{Load the alaska_dec (old ACToR - flex)data  from toxval sourcedb to toxval}
\usage{
toxval.load.alaska_dec(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{verbose}{If TRUE, output extra debug information}
}
\description{
Load the alaska_dec (old ACToR - flex)data  from toxval sourcedb to toxval
}
