% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.all.R
\name{toxval.load.all}
\alias{toxval.load.all}
\title{Load and process all information into ToxValDB. The entire process can be run with
one command: toxval.load.all(toxval.db=...,source.db=..., do.all=T)
It can also be run in stages, but needs to be run in the order of the do.X parameters
listed here. If any earlier step is run, all of the subsequent steps need to be rerun.}
\usage{
toxval.load.all(
  toxval.db,
  source.db,
  do.init = F,
  do.dict = F,
  do.load = F,
  do.load.chemical = F,
  do.dsstox = F,
  do.fix.species = F,
  do.fix.1 = F,
  do.fix.units = F,
  do.fix.final = F,
  do.all = F,
  do.post.load = F,
  do.hash = T,
  verbose = T,
  reset.hash = T,
  do.export = F
)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxvalsource database from which information is pulled.}

\item{do.init}{If True, clean out all of the database tables}

\item{do.dict}{If TRUE,  the initial dictionaries}

\item{do.load}{If TRUE, load all of the in vivo data}

\item{do.load.chemical}{If TRUE, load the chemical lists}

\item{do.dsstox, }{If TRUE, do the mapping to DSSTox to get the DSSTOx generic substance IDs}

\item{do.fix.species}{If TRUE, update all ofthe species information}

\item{do.fix.1}{If TRUE, clean up a series of database fields using dictionaries}

\item{do.fix.units}{If TRUE, fix the toxval_units field}

\item{do.all}{If TRUE, run all steps.}

\item{do.post.load}{If TRUE, to all steps past loading}

\item{do.hash}{If TRUE, set all of the hashes}

\item{verbose}{if TRUE, print disgnostic messages along the way}

\item{reset.hash}{If TRUE, reset the hashes instead of only setting values ='-'}

\item{do.export}{If TRUE, run all export scripts}

\item{fix.final}{If TRUE, run the final steps, including setting the record_id field}
}
\description{
Load and process all information into ToxValDB. The entire process can be run with
one command: toxval.load.all(toxval.db=...,source.db=..., do.all=T)
It can also be run in stages, but needs to be run in the order of the do.X parameters
listed here. If any earlier step is run, all of the subsequent steps need to be rerun.
}
