% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.atsdr.pfas.2021.R
\name{toxval.load.atsdr.pfas.2021}
\alias{toxval.load.atsdr.pfas.2021}
\title{Load new_atsdr_pfas_2021 from toxval_source to toxval}
\usage{
toxval.load.atsdr.pfas.2021(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load new_atsdr_pfas_2021 from toxval_source to toxval
}
