% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.cal_dph.R
\name{toxval.load.cal_dph}
\alias{toxval.load.cal_dph}
\title{Load the California DPH data (old ACToR - flex)data  from toxval sourcedb to toxval}
\usage{
toxval.load.cal_dph(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load the California DPH data (old ACToR - flex)data  from toxval sourcedb to toxval
}
