% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.cancer.R
\name{toxval.load.cancer}
\alias{toxval.load.cancer}
\title{prepare the cancer call data. The data comes form a series of files
../NIOSH/NIOSH_CARC_2018.xlsx
../IRIS/iris_cancer_call_2018-10-03.xlsx
../PPRTV_ORNL/PPRTV_ORNL cancer calls 2018-10-25.xlsx
../cancer_summary/cancer/NTP/NTP cancer clean.xlsx
../cancer_summary/cancer/IARC/IARC cancer 2018-10-29.xlsx
../cancer_summary/cancer/HealthCanada/HealthCanada_TRVs_2010_AppendixA v2.xlsx
../cancer_summary/cancer/EPA_OPP_CARC/EPA_CARC.xlsx
../cancer_summary/cancer/CalEPA/calepa_p65_cancer_only.xlsx}
\usage{
toxval.load.cancer(toxval.db)
}
\arguments{
\item{toxval.db}{The version of the database to use}
}
\description{
extract all of the chemicals with cancer slope factor or unit risk with appropriate units
}
