% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.chemical.list.R
\name{toxval.load.chemical.list}
\alias{toxval.load.chemical.list}
\title{Load The chemical lists to toxval. All Excel files in the folder ../chemicals/for_load/
are loaded. The must have the columns casrn, name, list_name}
\usage{
toxval.load.chemical.list(toxval.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{verbose}{If TRUE, print debugging messages}
}
\description{
Load The chemical lists to toxval. All Excel files in the folder ../chemicals/for_load/
are loaded. The must have the columns casrn, name, list_name
}
