% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.chemidplus.R
\name{toxval.load.chemidplus}
\alias{toxval.load.chemidplus}
\title{Load ChemID Plus Acute data data to toxval}
\usage{
toxval.load.chemidplus(toxval.db, source.db, log = F, do.init = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{log}{If TRUE, send output to a log file}

\item{do.init}{if TRUE, read the data in from the file and set up the matrix}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load ChemID Plus Acute data data to toxval
}
