% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.cosmos.R
\name{toxval.load.cosmos}
\alias{toxval.load.cosmos}
\title{Load teh COSMOS data from source to toxval}
\usage{
toxval.load.cosmos(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load teh COSMOS data from source to toxval
}
