% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.dod.ered.R
\name{toxval.load.dod.ered}
\alias{toxval.load.dod.ered}
\title{Load new_dod_table from toxval_source to toxval}
\usage{
toxval.load.dod.ered(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load new_dod_table from toxval_source to toxval
}
