% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.doe.ecorisk.R
\name{toxval.load.doe.ecorisk}
\alias{toxval.load.doe.ecorisk}
\title{Load the DOE ECORISK data (also called LANL) data from toxval_source to toxval}
\usage{
toxval.load.doe.ecorisk(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load the DOE ECORISK data (also called LANL) data from toxval_source to toxval
}
