% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.doe.pac.R
\name{toxval.load.doe.pac}
\alias{toxval.load.doe.pac}
\title{Load DOE Protective Action Criteria data from toxval_source to toxval}
\usage{
toxval.load.doe.pac(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load DOE Protective Action Criteria data from toxval_source to toxval
}
