% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.doe.R
\name{toxval.load.doe}
\alias{toxval.load.doe}
\title{Load new_doe_table and new_doe_benchmarks_table from toxval_source to toxval}
\usage{
toxval.load.doe(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load new_doe_table and new_doe_benchmarks_table from toxval_source to toxval
}
