% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.echa.echemportal.api.R
\name{toxval.load.echa.echemportal.api}
\alias{toxval.load.echa.echemportal.api}
\title{Load echa_echemportal_api from toxval_source to toxval}
\usage{
toxval.load.echa.echemportal.api(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load echa_echemportal_api from toxval_source to toxval
}
