% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.echa.iuclid.R
\name{toxval.load.echa.iuclid}
\alias{toxval.load.echa.iuclid}
\title{Load ECHA IUCLID data from source}
\usage{
toxval.load.echa.iuclid(toxval.db, source.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{verbose}{Whether the loaded rows should be printed to the console.}

\item{to.file}{If TRUE, output the data to a file for QC}

\item{do.read}{If TRUE read in the data file and store in a global}
}
\description{
Load ECHA IUCLID data from source
}
