% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.echa3.R
\name{toxval.load.echa3}
\alias{toxval.load.echa3}
\title{Load new_echa3 from toxval_source to toxval}
\usage{
toxval.load.echa3(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load new_echa3 from toxval_source to toxval
}
