% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.echa_iuclid.R
\name{toxval.load.echa_iuclid}
\alias{toxval.load.echa_iuclid}
\title{Loading the ECHA IUCLID data to toxval from toxval_source
This method is different from most because there are multiple tables (one per study
type) for this source}
\usage{
toxval.load.echa_iuclid(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{log}{If TRUE, send output to a log file}
}
\description{
Loading the ECHA IUCLID data to toxval from toxval_source
This method is different from most because there are multiple tables (one per study
type) for this source
}
