% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.ecotox.R
\name{toxval.load.ecotox}
\alias{toxval.load.ecotox}
\title{Load ECOTOX from the web services output to toxval}
\usage{
toxval.load.ecotox(toxval.db, source.db, log = F, do.load = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval source - used to manage chemicals}

\item{log}{If TRUE, send output to a log file}

\item{do.load}{If TRUE, load the data from the input file and put into a global variable}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load ECOTOX from the web services output to toxval
}
