% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.efsa.R
\name{toxval.load.efsa}
\alias{toxval.load.efsa}
\title{Load new_efsa from toxval_source to toxval}
\usage{
toxval.load.efsa(toxval.db, source.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load new_efsa from toxval_source to toxval
}
