% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.envirotox.R
\name{toxval.load.envirotox}
\alias{toxval.load.envirotox}
\title{Load original_envirotox from toxval_source to toxval}
\usage{
toxval.load.envirotox(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load original_envirotox from toxval_source to toxval
}
