% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.epa_aegl.R
\name{toxval.load.epa_aegl}
\alias{toxval.load.epa_aegl}
\title{Load the EPA AEGL (old ACToR - flex)data  from toxval sourcedb to toxval}
\usage{
toxval.load.epa_aegl(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load the EPA AEGL (old ACToR - flex)data  from toxval sourcedb to toxval
}
