% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.flex.R
\name{toxval.load.flex}
\alias{toxval.load.flex}
\title{Load the FLEX data (old ACToR data) from files to toxval. This will load all
Excel file in the folder ACToR replacements/}
\usage{
toxval.load.flex(toxval.db, verbose = F, only.new = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{verbose}{Whether the loaded rows should be printed to the console.}

\item{only.new}{if TRUE only files where the sources is not alrady in the database
will be loaded}
}
\description{
Load the FLEX data (old ACToR data) from files to toxval. This will load all
Excel file in the folder ACToR replacements/
}
