% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.generic.R
\name{toxval.load.generic}
\alias{toxval.load.generic}
\title{Generic structure for laoding to toxval from toxval_source}
\usage{
toxval.load.generic(toxvaldb, source.db, log = F)
}
\arguments{
\item{source.db}{The source database}

\item{log}{If TRUE, send output to a log file}

\item{toxval.db}{The database version to use}
}
\description{
Generic structure for laoding to toxval from toxval_source
}
