% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.genetox.all.R
\name{toxval.load.genetox.all}
\alias{toxval.load.genetox.all}
\title{Load the Genetox data from Grace}
\usage{
toxval.load.genetox.all(
  toxval.db,
  source.db,
  sys.date = "2021-09-10",
  verbose = FALSE
)
}
\arguments{
\item{toxval.db}{The database to use.}

\item{verbose}{If TRUE output debug information}

\item{do.read}{If TRUE, read in the DSSTox file}
}
\description{
Load the Genetox data from Grace
}
