% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.genetox.R
\name{toxval.load.genetox}
\alias{toxval.load.genetox}
\title{Load the Genetox data from Grace}
\usage{
toxval.load.genetox(toxval.db, verbose = F, do.read = T)
}
\arguments{
\item{toxval.db}{The database to use.}

\item{verbose}{If TRUE output debug information}

\item{do.read}{If TRUE, read in the DSSTox file}
}
\description{
Load the Genetox data from Grace
}
