% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.genetox_details.R
\name{toxval.load.genetox_details}
\alias{toxval.load.genetox_details}
\title{Load the Genetox data from Grace}
\usage{
toxval.load.genetox_details(toxval.db, verbose = F)
}
\arguments{
\item{toxval.db}{The database to use.}

\item{verbose}{if TRUE output debug information}
}
\description{
Load the Genetox data from Grace
}
