% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.hawc_pfas_150.R
\name{toxval.load.hawc.pfas.150}
\alias{toxval.load.hawc.pfas.150}
\title{Load HAWC PFAS 150 from toxval_source to toxval}
\usage{
toxval.load.hawc.pfas.150(toxval.db, source.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{verbose}{If TRUE, output extra debug information}
}
\description{
Load HAWC PFAS 150 from toxval_source to toxval
}
