% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.hawc_pfas_430.R
\name{toxval.load.hawc_pfas_430}
\alias{toxval.load.hawc_pfas_430}
\title{Load HAWC PFAS 430 from toxval_source to toxval}
\usage{
toxval.load.hawc_pfas_430(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load HAWC PFAS 430 from toxval_source to toxval
}
