% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.healthcanada.R
\name{toxval.load.healthcanada}
\alias{toxval.load.healthcanada}
\title{Load new_health_canada_table from toxval_source to toxval}
\usage{
toxval.load.healthcanada(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load new_health_canada_table from toxval_source to toxval
}
