% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.hpvis.R
\name{toxval.load.hpvis}
\alias{toxval.load.hpvis}
\title{Load HPVIS from toxval_source to toxval}
\usage{
toxval.load.hpvis(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source databse from which data should be loaded}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load HPVIS from toxval_source to toxval
}
