% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.initial.R
\name{toxval.load.initial}
\alias{toxval.load.initial}
\title{Delete the contents of the toxval database}
\usage{
toxval.load.initial(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval from which the data is deleted.}
}
\description{
Delete the contents of the toxval database
}
