% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.iris.R
\name{toxval.load.iris}
\alias{toxval.load.iris}
\title{Load new_iris_noncancer and new_iris_cancer from toxval_source to toxval}
\usage{
toxval.load.iris(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load new_iris_noncancer and new_iris_cancer from toxval_source to toxval
}
