% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.lanl.R
\name{toxval.load.lanl}
\alias{toxval.load.lanl}
\title{Load new_lanl_table from toxval_source to toxval}
\usage{
toxval.load.lanl(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load new_lanl_table from toxval_source to toxval
}
