% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.mass_mmcl.R
\name{toxval.load.mass_mmcl}
\alias{toxval.load.mass_mmcl}
\title{Load the mass_mmcl (old ACToR - flex)data  from toxval sourcedb to toxval}
\usage{
toxval.load.mass_mmcl(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load the mass_mmcl (old ACToR - flex)data  from toxval sourcedb to toxval
}
