% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.new_ecotox.R
\name{toxval.load.new_ecotox}
\alias{toxval.load.new_ecotox}
\title{Load ecotox data from datahub to toxval}
\usage{
toxval.load.new_ecotox(toxval.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load ecotox data from datahub to toxval
}
