% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.opp.R
\name{toxval.load.opp}
\alias{toxval.load.opp}
\title{Load opp from toxval_source to toxval}
\usage{
toxval.load.opp(toxval.db, source.db, verbose)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load opp from toxval_source to toxval
}
