% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.osha_air_limits.R
\name{toxval.load.osha_air_limits}
\alias{toxval.load.osha_air_limits}
\title{Load the osha_air_limits (old ACToR - flex)data  from toxval sourcedb to toxval}
\usage{
toxval.load.osha_air_limits(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load the osha_air_limits (old ACToR - flex)data  from toxval sourcedb to toxval
}
