% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.pfas_150_sem.R
\name{toxval.load.pfas.150.sem}
\alias{toxval.load.pfas.150.sem}
\title{Load pfas_150_sem from toxval_source to toxval}
\usage{
toxval.load.pfas.150.sem(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load pfas_150_sem from toxval_source to toxval
}
