% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.pfas_summary_pods.R
\name{toxval.load.pfas.summary.pods}
\alias{toxval.load.pfas.summary.pods}
\title{Load PFAS Summary PODs from toxval_source to toxval}
\usage{
toxval.load.pfas.summary.pods(toxval.db, source.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{verbose}{If TRUE, output extra debug information}
}
\description{
Load PFAS Summary PODs from toxval_source to toxval
}
