% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.pfas_150_sem.R
\name{toxval.load.pfas_150_sem}
\alias{toxval.load.pfas_150_sem}
\title{Load pfas_150_sem from toxval_source to toxval}
\usage{
toxval.load.pfas_150_sem(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load pfas_150_sem from toxval_source to toxval
}
