% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.pfas_summary_pods.R
\name{toxval.load.pfas_summary_pods}
\alias{toxval.load.pfas_summary_pods}
\title{Load PFAS Summary PODs from toxval_source to toxval}
\usage{
toxval.load.pfas_summary_pods(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load PFAS Summary PODs from toxval_source to toxval
}
