% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.postprocess.R
\name{toxval.load.postprocess}
\alias{toxval.load.postprocess}
\title{Do all of the post-processing steps for a source}
\usage{
toxval.load.postprocess(toxval.db, source.db, source, do.convert.units = T)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source}{The source name}

\item{do.convert.units}{If TRUE, convert units, mainly from ppm to mg/kg-day. This code is not debugged}

\item{sourcedb}{The source database name}
}
\description{
Do all of the post-processing steps for a source
}
