% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.pprtv.ncea.R
\name{toxval.load.pprtv.ncea}
\alias{toxval.load.pprtv.ncea}
\title{Load pprtv from dev_pprtv to toxval
There is a known bug here - some of the POD values are repeated because they produce
two kinds of RfD values (chronic and subchronic) - dealing with htis will require
some work}
\usage{
toxval.load.pprtv.ncea(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load pprtv from dev_pprtv to toxval
There is a known bug here - some of the POD values are repeated because they produce
two kinds of RfD values (chronic and subchronic) - dealing with htis will require
some work
}
