% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.pprtv.ornl.R
\name{toxval.load.pprtv.ornl}
\alias{toxval.load.pprtv.ornl}
\title{Load new_pprtv_ornl from toxval_source to toxval}
\usage{
toxval.load.pprtv.ornl(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source databse from which data should be loaded}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load new_pprtv_ornl from toxval_source to toxval
}
