% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.skin.eye.R
\name{toxval.load.skin.eye}
\alias{toxval.load.skin.eye}
\title{Load the Skin eye data}
\usage{
toxval.load.skin.eye(toxval.db, verbose = F)
}
\arguments{
\item{toxval.db}{Database version}

\item{verbose}{if TRUE, print diagnostic messages along the way}
}
\description{
Load the Skin eye data
}
