% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.source_chemical.R
\name{toxval.load.source_chemical}
\alias{toxval.load.source_chemical}
\title{Perform the DSSTox mapping}
\usage{
toxval.load.source_chemical(toxval.db, source.db, source = NULL, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}

\item{source.db}{The source database version}

\item{source}{The source to update for}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Perform the DSSTox mapping
}
