% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.species_ecotox.R
\name{toxval.load.species}
\alias{toxval.load.species}
\title{Load the species table and the species_id column in toxval}
\usage{
toxval.load.species(toxval.db, restart = F, date_string = "2022-05-25")
}
\arguments{
\item{toxval.db}{The version of the database to use}

\item{restart}{If TRUE, rest all of the species_id values in toxval}

\item{date.string}{Date suffix on the input species dictionary}
}
\description{
This function replaces fix.species
This function precedes toxvaldb.load.species
}
