% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.test.R
\name{toxval.load.test}
\alias{toxval.load.test}
\title{Load TEST data from toxval_source to toxval}
\usage{
toxval.load.test(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load TEST data from toxval_source to toxval
}
