% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.toxrefdb2.1.R
\name{toxval.load.toxrefdb2.1}
\alias{toxval.load.toxrefdb2.1}
\title{Load ToxRefdb data to toxval}
\usage{
toxval.load.toxrefdb2.1(toxval.db, source.db, log = F, do.init = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{log}{If TRUE, send output to a log file}

\item{do.init}{if TRUE, read the data in from the toxrefdb database and set up the matrix}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load ToxRefdb data to toxval
}
