% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.tri.R
\name{toxval.load.tri}
\alias{toxval.load.tri}
\title{Load TRI data data to toxval}
\usage{
toxval.load.tri(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{log}{If TRUE, send output to a log file}

\item{verbose}{Whether the loaded rows should be printed to the console.}

\item{do.init}{if TRUE, read the data in from the file and set up the matrix}
}
\description{
Load TRI data data to toxval
}
