% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.ut_hb.R
\name{toxval.load.ut_hb}
\alias{toxval.load.ut_hb}
\title{Load teh uterotophic and Hershberger data}
\usage{
toxval.load.ut_hb(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{log}{If TRUE, send output to a log file}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load teh uterotophic and Hershberger data
}
