% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.qc.step.1.R
\name{toxval.qc.step.1}
\alias{toxval.qc.step.1}
\title{do an initial QC of the data by comparing the current database to an old one}
\usage{
toxval.qc.step.1(db.new = "res_toxval_v92", db.old = "dev_toxval_v9_1")
}
\arguments{
\item{db.new}{The new database version (toxval) for the comparison}

\item{db.old}{= The old database version for the comparison}
}
\description{
do an initial QC of the data by comparing the current database to an old one
}
