% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.set.mw.R
\name{toxval.set.mw}
\alias{toxval.set.mw}
\title{Set the molecular weight in the toxval table, for use in unit conversions}
\usage{
toxval.set.mw(toxval.db, source = NULL)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source}{The source}
}
\description{
Set the molecular weight in the toxval table, for use in unit conversions
}
