% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval_source_add_timestamps.R
\name{toxval.source.add.timestamps}
\alias{toxval.source.add.timestamps}
\title{toxval.source.add.timestamps}
\usage{
toxval.source.add.timestamps(source.db)
}
\arguments{
\item{source.db}{The version of toxval source database to use.}
}
\value{
None. SQL statements are executed to add timestamp columns.
}
\description{
get chemical info from source db tables
}
