% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.source.import.dedup.R
\name{toxval.source.import.dedup}
\alias{toxval.source.import.dedup}
\title{toxval.source.import.dedup}
\usage{
toxval.source.import.dedup(
  res,
  dedup_fields = NULL,
  hashing_cols = toxval.config()$hashing_cols,
  delim = " |::| "
)
}
\arguments{
\item{res}{dataframe containing the source data to dedup}

\item{dedup_fields}{vector containing field names to dedup, Default: NULL (all fields but hashing cols)}

\item{hashing_cols}{vector containing field names of hashing columns, Default: toxval.config()$hashing_cols}

\item{delim}{string used to separate collapsed values, Default: ' |::| '}
}
\value{
dataframe containing deduped source data
}
\description{
Perform deduping on data before it is sent to toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{na_if}}, \code{\link[dplyr]{distinct}}
}
