% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.source.map.chemicals.combined.R
\name{toxval.source.map.chemicals.combined}
\alias{toxval.source.map.chemicals.combined}
\title{toxval.source.map.chemicals.combined}
\usage{
toxval.source.map.chemicals.combined(
  source.db,
  input.path,
  curated.path,
  match.raw = FALSE
)
}
\arguments{
\item{source.db}{The version of toxval source database to use.}

\item{input.path}{Path to folder with original chemical lists}

\item{curated.path}{Path to folder with curated chemical lists}

\item{match.raw}{Boolean whether to match by raw name/casrn values (Default FALSE)}
}
\value{
None. SQL statements are executed.
}
\description{
push ChemReg curated chemicals to toxval_source db tables
}
