% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.source.map.chemicals.R
\name{toxval.source.map.chemicals}
\alias{toxval.source.map.chemicals}
\title{toxval.source.map.chemicals}
\usage{
toxval.source.map.chemicals(source.db, input.path, curated.path)
}
\arguments{
\item{source.db}{The version of toxval source database to use.}

\item{input.path}{Path to folder with original chemical lists}

\item{curated.path}{Path to folder with curated chemical lists}
}
\value{
None. SQL statements are executed.
}
\description{
push ChemReg curated chemicals to toxval_source db tables
}
