% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.source_push_mapped_chemicals.R
\name{toxval.source_push_mapped_chemicals}
\alias{toxval.source_push_mapped_chemicals}
\title{toxval.source_push_mapped_chemicals}
\usage{
toxval.source_push_mapped_chemicals(
  db,
  source.index,
  curated.path,
  ignore.curation.dups = TRUE,
  match.chemical.id = FALSE,
  reset.mapping = FALSE,
  bulk.push = TRUE
)
}
\arguments{
\item{db}{The version of toxval source database to use.}

\item{source.index}{The source chemical index. Can be full or just numeric (ex. ToxVal00001 vs. 00001)}

\item{curated.path}{Input path to the folder directory with expected subdirectories of #' 'BIN Files', 'DSSTox Files', and 'jira_chemical_files'}

\item{ignore.curation.dups}{Boolean whether to match with any curated records flagged as "unresolved duplicates" (Default TRUE)}

\item{match.chemical.id}{Boolean whether to match by provided chemical_id external identifier (Default FALSE)}

\item{reset.mapping}{Boolean whether to reset chemical mappings in source_chemical table of database}

\item{bulk.push}{Boolean whether to bulk push updates, or one at a time. Default is TRUE}
}
\value{
None. Update SQL statements are executed.
}
\description{
Orchestrates the push of mapped chemical information to the selected toxval_source database.
Uses the map_curated_chemicals() helper function to generate mapped input.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{rename}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{bind}}
}
