% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.summary.stats.R
\name{toxval.summary.stats}
\alias{toxval.summary.stats}
\title{Generate summary statistics on the toxval database}
\usage{
toxval.summary.stats(toxval.db)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}
}
\description{
Generate summary statistics on the toxval database
}
