% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval_source.hash.and.load.R
\name{toxval_source.hash.and.load}
\alias{toxval_source.hash.and.load}
\title{Add the hash key to the source tables and add the new rows}
\usage{
toxval_source.hash.and.load(
  db = "dev_toxval_source_v5",
  source,
  table,
  do.reset = F,
  do.insert = F,
  res
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{source}{Name of the source}

\item{table}{Name of the database table}

\item{do.reset}{If TRUE, delete data from the database for this source before
inserting new data. Default FALSE}

\item{do.insert}{If TRUE, insert data into the database, default TRUE}

\item{res}{The data frame to be processed}
}
\description{
Add the hash key to the source tables and add the new rows
}
