% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.clowder.docs.R
\name{update.clowder.docs}
\alias{update.clowder.docs}
\title{Update the clowder links in toxval_source}
\usage{
\method{update}{clowder.docs}(toxval.db, source.db, dir = "updates_2022-08-09")
}
\arguments{
\item{dir}{The directory where the update files sit}

\item{db}{Database version of the source database}
}
\description{
Update the clowder links in toxval_source
}
