% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.source.clowder.id.R
\name{update.source.clowder.id}
\alias{update.source.clowder.id}
\title{Create audit table and add BEFORE UPDATE audit triggers to source_* tables}
\usage{
\method{update}{source.clowder.id}(
  src_tbl,
  source,
  map_file = NULL,
  db,
  do.halt = FALSE,
  verbose = FALSE,
  reset = FALSE
)
}
\arguments{
\item{src_tbl}{the name of the ToxVal source table to update}

\item{source}{the name of the ToxVal source (different from src_tbl)}

\item{map_file}{the file path to a Clowder document map}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}

\item{reset}{if TRUE, fully replace/update Clowder ID values, even if ID values exist}
}
\description{
Create audit table and add BEFORE UPDATE audit triggers to source_* tables
}
