% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem.check.v2.R
\name{chem.check.v2}
\alias{chem.check.v2}
\title{Check the chemicals from a file
Names with special characters are cleaned and trimmed
CASRN are fixed (dashes put in, trimmed) and check sums are calculated
The output is sent to a file called chemcheck.xlsx in the source data file
One option for using this is to edit the source file until no errors are found}
\usage{
chem.check.v2(res0, source = NULL, verbose = F)
}
\arguments{
\item{res0}{The data frame in which chemicals names and CASRN will be replaced}

\item{source}{The source to be processed. If source=NULL, process all sources}

\item{verbose}{If TRUE, print diagnostic messages}
}
\value{
Return a list with fixed CASRN and name and flags indicating if fixes were made:
res0=res0,name.OK=name.OK,casrn.OK=casrn.OK,checksum.OK=checksum.OK
}
\description{
Check the chemicals from a file
Names with special characters are cleaned and trimmed
CASRN are fixed (dashes put in, trimmed) and check sums are calculated
The output is sent to a file called chemcheck.xlsx in the source data file
One option for using this is to edit the source file until no errors are found
}
