% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.last.character.R
\name{clean.last.character}
\alias{clean.last.character}
\title{Clean unneeded characters from the end of a string}
\usage{
clean.last.character(x)
}
\arguments{
\item{x}{String to be cleaned}
}
\value{
The cleaned string
}
\description{
Clean unneeded characters from the end of a string
}
