% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source_table_SQL.R
\name{create_source_table_SQL}
\alias{create_source_table_SQL}
\title{FUNCTION_TITLE}
\usage{
create_source_table_SQL(
  source,
  res,
  src_version,
  db,
  do.halt = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{source}{name of the source being processed}

\item{res}{input dataframe of source data}

\item{src_version}{Version date of the source}

\item{db}{PARAM_DESCRIPTION}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Input source data is used to generate the SQL for the source's toxval_source table.
SQL is based off a generic SQL file
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_trim}}
}
