% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAT.manual.pipe.source.audit.R
\name{DAT.manual.pipe.source.audit}
\alias{DAT.manual.pipe.source.audit}
\title{Processes manually QC'd (sans-DAT) QC audit information into database}
\usage{
DAT.manual.pipe.source.audit(source, db, live_df, qc_user = "Evelyn Rowan")
}
\arguments{
\item{source}{name of ToxVal source table audit information is associated with}

\item{db}{the name of the database}

\item{live_df}{a filepath to the DAT live data to push to the 'source' table}

\item{qc_user}{The name of the user who completed the QC}
}
\description{
Processes manually QC'd (sans-DAT) QC audit information into database
}
