% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAT.pipe.source.audit.R
\name{DAT.pipe.source.audit}
\alias{DAT.pipe.source.audit}
\title{DAT.pipe.source.audit}
\usage{
DAT.pipe.source.audit(
  source_table,
  db,
  live_df,
  audit_df,
  hashing_type = "vectorized"
)
}
\arguments{
\item{source_table}{name of ToxVal source table audit information is associated with}

\item{db}{the name of the database}

\item{live_df}{a filepath to the DAT live data to push to the 'source' table}

\item{audit_df}{a filepath to the DAT audit data to push to source_audit #'}

\item{hashing_type}{character string of 'vectorized' or 'base' representing which
source_hash generation approach to use for the data based on how the original source was hashed.}
}
\value{
None
}
\description{
Processes DAT QC audit information into database
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{filter}}
 \code{\link[tidyr]{reexports}}
 \code{\link[writexl]{write_xlsx}}
}
