% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.casrn.R
\name{fix.casrn}
\alias{fix.casrn}
\title{fix.casrn}
\usage{
fix.casrn(casrn, cname = "", verbose = FALSE)
}
\arguments{
\item{casrn}{Input CASRN to be fixed}

\item{cname}{An optional chemical name}

\item{verbose}{if TRUE, print the input values}
}
\value{
the fixed CASRN
}
\description{
Fix a CASRN that has one of several problems
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 fix.casrn("107028")
 # Expected output "107-02-8"
 }
}
}
\seealso{
\code{\link[tidyr]{reexports}}
}
