% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.priority_id.by.source.R
\name{fix.priority_id.by.source}
\alias{fix.priority_id.by.source}
\title{Fix the priority_id in the toxval table based on source}
\usage{
fix.priority_id.by.source(toxval.db, source)
}
\arguments{
\item{toxval.db}{The version of toxvaldb to use.}
}
\description{
Fix the priority_id in the toxval table based on source
}
