% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.strain.v2.R
\name{fix.strain.v2}
\alias{fix.strain.v2}
\title{Set the strain information in toxval}
\usage{
fix.strain.v2(toxval.db, source = NULL, date_string = "2022-08-15")
}
\arguments{
\item{toxval.db}{The version of the database to use}

\item{source}{The source to be fixed. If NULL, fix for all sources}

\item{date_string}{The date of the latest dictionary version}
}
\description{
Set the strain information in toxval
}
