% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.dictionary.R
\name{import.dictionary}
\alias{import.dictionary}
\title{import the toxval and toxval_type dictionaries}
\usage{
import.dictionary(toxval.db)
}
\arguments{
\item{toxval.db}{The name of the database}
}
\description{
import the toxval and toxval_type dictionaries
}
