% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.driver.R
\name{import.driver}
\alias{import.driver}
\title{FUNCTION_TITLE}
\usage{
import.driver(
  db = "res_toxval_source_v5",
  chem.check.halt = FALSE,
  do.clean = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{PARAM_DESCRIPTION, Default: FALSE}

\item{do.clean}{If TRUE, delte data from all tables before reloading}

\item{chem.chek.halt}{If TRUE and there are bad chemical names or casrn, #' stop to look at the results in indir/chemcheck.xlsx}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function to run all import scripts to fill toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
