% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atsdr_pfas_2021_source.R
\name{import_atsdr_pfas_2021_source}
\alias{import_atsdr_pfas_2021_source}
\title{import_atsdr_pfas_2021_source}
\usage{
import_atsdr_pfas_2021_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is pushed to ToxVal
}
\description{
Load ATSDR PFAS 2021 data to toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[janitor]{remove_empty}}
 \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}}
 \code{\link[tidyr]{separate_rows}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{drop_na}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_extract}}, \code{\link[stringr]{str_replace}}
 \code{\link[tibble]{enframe}}
 \code{\link[tidyselect]{all_of}}
}
