% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atsdr_pfas_2021_source.R
\name{import_atsdr_pfas_2021_source}
\alias{import_atsdr_pfas_2021_source}
\title{Load ATSDR PFAS 2021 Source into toxval_source}
\usage{
import_atsdr_pfas_2021_source(db, chem.check.halt = F)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}

\item{indir}{The path for all the input xlsx files ./atsdr_pfas_2021/atsdr_pfas_2021_files}
}
\description{
Load ATSDR PFAS 2021 Source into toxval_source
}
