% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atsdr_pfas_source.R
\name{import_atsdr_pfas_source}
\alias{import_atsdr_pfas_source}
\title{Load ATSDR PFAS Source files into toxval_source}
\usage{
import_atsdr_pfas_source(
  db,
  infile1 = "ATSDR_Perfluoroalkyls_Inhalation.xlsx",
  infile2 = "ATSDR_Perfluoroalkyls_Oral.xlsx",
  infile3 = "ATSDR_PFOA_Inhalation.xlsx",
  infile4 = ".ATSDR_PFOA_Oral.xlsx",
  infile5 = "ATSDR_PFOS_Oral.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile1}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_Perfluoroalkyls_Inhalation.xlsx}

\item{infile2}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_Perfluoroalkyls_Oral.xlsx}

\item{infile3}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_PFOA_Inhalation.xlsx}

\item{infile4}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_PFOA_Oral.xlsx}

\item{infile5}{The input file ./atsdr_pfas/atsdr_pfas_files/ATSDR_PFOS_Oral.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load ATSDR PFAS Source files into toxval_source
}
