% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_atsdr_source.R
\name{import_atsdr_source}
\alias{import_atsdr_source}
\title{FUNCTION_TITLE}
\usage{
import_atsdr_source(
  db,
  infile = "ATSDR_MRLs_2020_Sept2020_Temp.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./atsdr/atsdr_files/ATSDR_MRLs_2020_Sept2020_Temp.xls}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load atsdr Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[janitor]{excel_numeric_to_date}}
}
