% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_caloehha_source.R
\name{import_caloehha_source}
\alias{import_caloehha_source}
\title{FUNCTION_TITLE}
\usage{
import_caloehha_source(
  db,
  infile = "OEHHA-chemicals_2022-06-22T13-42-44.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ="../caloehha/caloehha_files/OEHHA-chemicals_2018-10-30T08-50-47.xlsx",}

\item{chem.check.halt}{If TRUE and there are problems with chemicals CASRN checks, halt the program}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load caloehha Source file into toxval_source
The raw data can be exported as an Excel sheet from the web site
https://oehha.ca.gov/chemicals, selecting the link "Export database as .CSV file"

This method parses that file and prepares for loading into toxval source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
