% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chiu_source.R
\name{import_chiu_source}
\alias{import_chiu_source}
\title{import_chiu_source}
\usage{
import_chiu_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn, #' stop to look at the results in indir/chemcheck.xlsx}

\item{infile}{The input file ./chiu/chiu_files/RfD_HDMI_mc_results.csv}
}
\value{
None; data is pushed to toxval_source
}
\description{
Load chiu Source into toxval_source, only HDMI values
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{select}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{distinct}}
 \code{\link[tidyr]{drop_na}}
 \code{\link[stringr]{str_remove}}, \code{\link[stringr]{modifiers}}, \code{\link[stringr]{str_trim}}
}
