% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_chiu_source.R
\name{import_chiu_source}
\alias{import_chiu_source}
\title{FUNCTION_TITLE}
\usage{
import_chiu_source(
  db,
  infile = "Full_RfD_databaseQAed-FINAL.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./chiu/chiu_files/Full_RfD_databaseQAed-FINAL.xlsx}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn, #' stop to look at the results in indir/chemcheck.xlsx}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
#' Load chiu Source into dev_toxval_source_v3.
Data from the Chiu et al. paper on RfD values

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
