% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_copper_source.R
\name{import_copper_source}
\alias{import_copper_source}
\title{Load copper manufacturers Source into toxval_source}
\usage{
import_copper_source(
  db,
  infile = "Copper Data Entry - Final.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./copper/copper_files/Copper Data Entry - Final.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load copper manufacturers Source into toxval_source
}
