% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cosmos_source.R
\name{import_cosmos_source}
\alias{import_cosmos_source}
\title{Load cosmos Source files into dev_toxval_source_v4.}
\usage{
import_cosmos_source(toxval.db, infile1, infile2)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile1}{The input file ./cosmos/cosmos_files/COSMOS_DB_v1_export_2016_04_02_study_data.xlsx}

\item{infile2}{The input file ./cosmos/cosmos_files/COSMOS_DB_v1_export_2016_04_02_cosmetics_inventory.xlsx}
}
\description{
Load cosmos Source files into dev_toxval_source_v4.
}
