% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cosmos_source.R
\name{import_cosmos_source}
\alias{import_cosmos_source}
\title{Load cosmos Source files into toxval_source}
\usage{
import_cosmos_source(
  db,
  infile1 = "COSMOS_DB_v1_export_2016_04_02_study_data.xlsx",
  infile2 = "COSMOS_DB_v1_export_2016_04_02_cosmetics_inventory.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile1}{The input file ./cosmos/cosmos_files/COSMOS_DB_v1_export_2016_04_02_study_data.xlsx}

\item{infile2}{The input file ./cosmos/cosmos_files/COSMOS_DB_v1_export_2016_04_02_cosmetics_inventory.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load cosmos Source files into toxval_source
}
