% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dod_ered_source.R
\name{import_dod_ered_source}
\alias{import_dod_ered_source}
\title{Load dod Source into dev_toxval_source_v2.}
\usage{
import_dod_ered_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./dod/dod_files/USACE_ERDC_ERED_database_12_07_2018.xlsx}
}
\description{
Load dod Source into dev_toxval_source_v2.
}
