% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dod_ered_source.R
\name{import_dod_ered_source}
\alias{import_dod_ered_source}
\title{Load dod Source into toxval_source}
\usage{
import_dod_ered_source(
  db,
  infile = "USACE_ERDC_ERED_database_12_07_2018.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./dod/dod_files/USACE_ERDC_ERED_database_12_07_2018.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load dod Source into toxval_source
}
