% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dod_source.R
\name{import_dod_source}
\alias{import_dod_source}
\title{Load DOD MEG to toxval_source. The file to be loaded are in ./dod/dod_files}
\usage{
import_dod_source(db, chem.check.halt = F)
}
\arguments{
\item{db}{The version of toxval_source into which the tables are loaded.}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load DOD MEG to toxval_source. The file to be loaded are in ./dod/dod_files
}
