% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doe_benchmarks_source.R
\name{import_doe_benchmarks_source}
\alias{import_doe_benchmarks_source}
\title{Load doe_benchmarks Source into dev_toxval_source_v2.}
\usage{
import_doe_benchmarks_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./doe_benchmarks/doe_benchmarks_files/DOE_Wildlife_Benchmarks_1996.xlsx}
}
\description{
Load doe_benchmarks Source into dev_toxval_source_v2.
}
