% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doe_benchmarks_source.R
\name{import_doe_benchmarks_source}
\alias{import_doe_benchmarks_source}
\title{import_doe_benchmarks_source}
\usage{
import_doe_benchmarks_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is pushed to toxval_source
}
\description{
Load DOE Wildlife Benchmarks data into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{row_number}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{distinct}}
 \code{\link[tidyselect]{starts_with}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_extract}}
 \code{\link[tidyr]{reexports}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{separate}}
}
