% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doe_benchmarks_source.R
\name{import_doe_benchmarks_source}
\alias{import_doe_benchmarks_source}
\title{FUNCTION_TITLE}
\usage{
import_doe_benchmarks_source(
  db,
  infile = "DOE_Wildlife_Benchmarks_1996.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./doe_benchmarks/doe_benchmarks_files/DOE_Wildlife_Benchmarks_1996.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load doe_benchmarks Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
