% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_doe_pac_source.R, R/import_doe_source.R
\name{import_doe_source}
\alias{import_doe_source}
\title{Load DOE Source into toxval_source}
\usage{
import_doe_source(toxval.db, infile)

import_doe_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./doe/doe_files/Revision_29.xlsx}

\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load DOE Source into toxval_source

Load doe Source into dev_toxval_source_v4.
}
