% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_echa_echemportal_api_source.R
\name{import_echa_echemportal_api_source}
\alias{import_echa_echemportal_api_source}
\title{Load ECHA echemportal api Source into dev_toxval_source_v4.}
\usage{
import_echa_echemportal_api_source(toxval.db, filepath)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{filepath}{The path for all the input xlsx files ./echa_echemportal_api/echa_echemportal_api_files}
}
\description{
Load ECHA echemportal api Source into dev_toxval_source_v4.
}
