% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_echa_echemportal_source.R
\name{import_echa_echemportal_source}
\alias{import_echa_echemportal_source}
\title{FUNCTION_TITLE}
\usage{
import_echa_echemportal_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./echa_echemportal/echa_echemportal_files/eChemPortal mammalian data 2020 step 3.xlsx ,build from echemportal.prep.v2.step3.R}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load ECHA echemportal 2020 Source into dev_toxval_source_v4.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[stringr]{str_extract}}, \code{\link[stringr]{modifiers}}
}
