% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_echa_iuclid_source.R
\name{import_echa_iuclid_source}
\alias{import_echa_iuclid_source}
\title{FUNCTION_TITLE}
\usage{
import_echa_iuclid_source(toxval.db, infile, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./echa_iuclid/echa_iuclid_files/echa_iuclid_v8.xlsx}

\item{verbose}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load ECHA IUCLID Source into dev_toxval_source_v4.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
