% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_efsa2_source.R
\name{import_efsa2_source}
\alias{import_efsa2_source}
\title{FUNCTION_TITLE}
\usage{
import_efsa2_source(
  db,
  infile = "EFSA_combined_new 2022-07-19.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./efsa2/efsa2_files/merge2/EFSA_combined_new.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load efsa2 Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[utils]{type.convert}}
}
