% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_envirotox_source.R
\name{import_envirotox_source}
\alias{import_envirotox_source}
\title{Load EnviroTox.V2 Source data into toxval_source}
\usage{
import_envirotox_source(
  db,
  infile = "envirotox_taxonomy clean casrn.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{infile}{The input file ./envirotox/envirotox_files/envirotox_taxonomy.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load EnviroTox.V2 Source data into toxval_source
}
