% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_flex_source.R
\name{import_flex_source}
\alias{import_flex_source}
\title{Load the FLEX data (old ACToR data) from files to toxval source. This will load all
Excel file in the folder ACToR replacements/}
\usage{
import_flex_source(
  db,
  filepath = "ACToR replacements",
  verbose = F,
  chem.check.halt = F,
  do.clean = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the tables are loaded.}

\item{filepath}{The path for all the input xlsx files ./ACToR replacements}

\item{verbose}{Whether the loaded rows should be printed to the console.}

\item{chem.check.halt}{If TRUE and there are problems with chemicals CASRN checks, halt the program}

\item{do.clean}{If true, remove data for these sources before reloading}
}
\description{
Load the FLEX data (old ACToR data) from files to toxval source. This will load all
Excel file in the folder ACToR replacements/
}
