% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_generic_source.R
\name{import_generic_source}
\alias{import_generic_source}
\title{A generic tmpalte for adding data to toxval_source for a new source}
\usage{
import_generic_source(db, chem.check.halt = F)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}
}
\description{
A generic tmpalte for adding data to toxval_source for a new source
}
