% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_generic_source.R,
%   R/import_pprtv_ncea_source.R, R/import_source_fda_cedi.R,
%   R/import_source_usgs_hbsl.R, R/import_source_who_ipcs.R,
%   R/import_source_who_jecfa_adi.R
\name{import_generic_source}
\alias{import_generic_source}
\alias{import_pprtv_ncea_source}
\alias{import_source_fda_cedi}
\alias{import_source_usgs_hbsl}
\alias{import_who_ipcs}
\alias{import_source_who_jecfa_adi}
\title{FUNCTION_TITLE}
\usage{
import_generic_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_pprtv_ncea_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_source_fda_cedi(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_source_usgs_hbsl(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_who_ipcs(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_source_who_jecfa_adi(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
OUTPUT_DESCRIPTION

None; data is pushed to ToxVal

OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION

None. Data is processed into the toxval_source database

OUTPUT_DESCRIPTION
}
\description{
A generic template for adding data to toxval_source for a new source

Load PPRTV (NCEA) data to toxval_source

Import FDA CEDI data into source_fda_cedi

Import USGS HBSL data into toxval_source

Import of WHO IPCS data

Import of WHO JECFA ADI data
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[purrr]{map}}
 \code{\link[readxl]{read_excel}}
 \code{\link[readr]{read_delim}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{na_if}}, \code{\link[dplyr]{across}}
 \code{\link[tidyselect]{all_of}}
 \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}, \code{\link[tidyr]{drop_na}}
 \code{\link[stringr]{str_trim}}

\code{\link[readr]{read_csv}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}
}
