% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_generic_source.R,
%   R/import_source_hess.R, R/import_source_ntp_pfas.R,
%   R/import_source_who_ipcs.R, R/import_source_who_jecfa_adi.R
\name{import_generic_source}
\alias{import_generic_source}
\alias{import_source_hess}
\alias{import_source_ntp_pfas}
\alias{import_who_ipcs}
\alias{import_source_who_jecfa_adi}
\title{FUNCTION_TITLE}
\usage{
import_generic_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_source_hess(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_source_ntp_pfas(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_who_ipcs(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)

import_source_who_jecfa_adi(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
OUTPUT_DESCRIPTION

None. Data is processed into the database

OUTPUT_DESCRIPTION

None. Data is processed into the toxval_source database

OUTPUT_DESCRIPTION
}
\description{
A generic template for adding data to toxval_source for a new source

Import of HESS data source

A functio for adding source NTP PFAS data to toxval_source

Import of WHO IPCS data

Import of WHO JECFA ADI data
}
\details{
DETAILS

DETAILS

DETAILS

DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}

\code{\link[readxl]{read_excel}}
 \code{\link[stringr]{str_trim}}
}
