% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hawc_pfas_150_source.R
\name{import_hawc_pfas_150_source}
\alias{import_hawc_pfas_150_source}
\title{Load HAWC PFAS 150 Source into dev_toxval_source_v4.}
\usage{
import_hawc_pfas_150_source(toxval.db, infile1, infile2, infile3)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile1}{The input file ./hawc_pfas/hawc_pfas_files/hawc_pfas_150_raw3.xlsx , extracted
from https://hawcprd.epa.gov , assessment name - PFAS 150 (2021) and assessment id - 100500085. 
Data extraction using HawcClient and extraction script hawc_pfas_150.py}

\item{infile2}{The input file ./hawc_pfas/hawc_pfas_files/hawc_pfas_150_doses3.xlsx}

\item{infile3}{The input file ./hawc_pfas/hawc_pfas_files/hawc_pfas_150_groups3.xlsx}
}
\description{
Load HAWC PFAS 150 Source into dev_toxval_source_v4.
}
