% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hawc_pfas_430_source.R
\name{import_hawc_pfas_430_source}
\alias{import_hawc_pfas_430_source}
\title{Load HAWC PFAS 430 Source into toxval_source}
\usage{
import_hawc_pfas_430_source(
  db,
  infile1 = "hawc_pfas_430_raw3.xlsx",
  infile2 = "hawc_pfas_430_doses3.xlsx",
  infile3 = "hawc_pfas_430_groups3.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile1}{The input file ./hawc_pfas_430/hawc_pfas_430_files/hawc_pfas_430_raw3.xlsx , extracted
from https://hawcprd.epa.gov , assessment name - PFAS 430 (2020) and assessment id - 100500256.
Data extraction using HawcClient and extraction script hawc_pfas_430.py}

\item{infile2}{The input file ./hawc_pfas_430/hawc_pfas_430_files/hawc_pfas_430_doses3.xlsx}

\item{infile3}{The input file ./hawc_pfas_430/hawc_pfas_430_files/hawc_pfas_430_groups3.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load HAWC PFAS 430 Source into toxval_source
}
