% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_epa_hawc_source.R,
%   R/import_hawc_pfas_source.R
\name{import_epa_hawc_source}
\alias{import_epa_hawc_source}
\alias{import_hawc_pfas_source}
\title{import_epa_hawc_source}
\usage{
import_epa_hawc_source(df_list = NULL)

import_hawc_pfas_source(
  db,
  hawc_num = NULL,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{df_list}{List of dataframes of "raw", "doses" and "groups" from HAWC API for an assessment ID.}

\item{db}{The version of toxval_source into which the source is loaded.}

\item{hawc_num}{The HAWC number being processed (e.g. 150, 430)}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
Returns processed HAWC data by assessment ID.

None; data is added to toxval_source
}
\description{
Prepare EPA HAWC data from input dataframe list.

Load HAWC PFAS data into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{case_when}}
 \code{\link[tidyselect]{all_of}}
 \code{\link[tidyr]{pivot_wider}}, \code{\link[tidyr]{unite}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{separate_rows}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_detect}}, \code{\link[stringr]{str_extract}}
 \code{\link[purrr]{map2}}
 \code{\link[digest]{digest}}

\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{case_when}}
 \code{\link[tidyselect]{all_of}}
 \code{\link[tidyr]{pivot_wider}}, \code{\link[tidyr]{unite}}, \code{\link[tidyr]{pivot_longer}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{drop_na}}, \code{\link[tidyr]{separate_rows}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_detect}}, \code{\link[stringr]{str_extract}}
 \code{\link[purrr]{map2}}
 \code{\link[digest]{digest}}
}
