% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hawc_source.R
\name{import_hawc_source}
\alias{import_hawc_source}
\title{import_hawc_source}
\usage{
import_hawc_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is pushed to toxval_source
}
\description{
Load HAWC Project data into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{getSheetNames}}, \code{\link[openxlsx]{read.xlsx}}
 \code{\link[dplyr]{select}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{context}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{na_if}}
 \code{\link[tidyr]{pivot_wider}}, \code{\link[tidyr]{unite}}, \code{\link[tidyr]{reexports}}, \code{\link[tidyr]{drop_na}}
 \code{\link[generics]{setops}}
 \code{\link[digest]{digest}}
 \code{\link[stringr]{str_extract}}, \code{\link[stringr]{str_trim}}
}
