% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hawc_source.R
\name{import_hawc_source}
\alias{import_hawc_source}
\title{Load HAWC Source into dev_toxval_source_v3.}
\usage{
import_hawc_source(toxval.db, infile1, infile2)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile1}{The input file ./hawc/hawc_files/hawc_original_12_06_21.xlsx}

\item{infile2}{The input file ./hawc/hawc_files/dose_dict.xlsx}
}
\description{
Load HAWC Source into dev_toxval_source_v3.
}
