% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_health_canada_source.R
\name{import_health_canada_source}
\alias{import_health_canada_source}
\title{FUNCTION_TITLE}
\usage{
import_health_canada_source(
  db,
  infile = "HealthCanada_TRVs_2010_AppendixA v2.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{infile}{The input file ./health_canada/health_canada_files/HealthCanada_TRVs_2010_AppendixA v2.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load Health Canada Source Info into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
