% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_health_canada_source.R
\name{import_health_canada_source}
\alias{import_health_canada_source}
\title{Load health_canada Source Info into dev_toxval_source_v2.}
\usage{
import_health_canada_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{infile}{The input file ./health_canada/health_canada_files/HealthCanada_TRVs_2010_AppendixA v2.xlsx}
}
\description{
Load health_canada Source Info into dev_toxval_source_v2.
}
