% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_heast_source.R
\name{import_heast_source}
\alias{import_heast_source}
\title{Load heast Source into dev_toxval_source_v2.}
\usage{
import_heast_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./heast/heast_files/EPA_HEAST_Table1_ORNL for loading.xlsx}
}
\description{
Load heast Source into dev_toxval_source_v2.
}
