% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hess_source.R
\name{import_hess_source}
\alias{import_hess_source}
\title{FUNCTION_TITLE}
\usage{
import_hess_source(
  db,
  infile1 = "hess_6_16_21.xlsx",
  infile2 = "hess_record_urls_from_clowder.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile1}{The input file ./hess/hess_files/hess_6_16_21.csv, extracted by Risa Sayre(SCDCD)}

\item{infile2}{The input file ./hess/hess_files/hess_record_urls_from_clowder.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load HESS Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[utils]{type.convert}}
 \code{\link[tidyr]{unite}}
 \code{\link[stringr]{str_replace}}
}
