% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hpvis_source.R
\name{import_hpvis_source}
\alias{import_hpvis_source}
\title{Load HPVIS Source Info into toxval_source}
\usage{
import_hpvis_source(db, filepath = "hpvis/hpvis_files", chem.check.halt = T)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{filepath}{The path for all the input xlsx files ./hpvis/hpvis_files}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load HPVIS Source Info into toxval_source
}
