% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hpvis_source.R
\name{import_hpvis_source}
\alias{import_hpvis_source}
\title{FUNCTION_TITLE}
\usage{
import_hpvis_source(db, filepath = "hpvis/hpvis_files", chem.check.halt = T)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{filepath}{The path for all the input xlsx files ./hpvis/hpvis_files}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load HPVIS Source Info into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[stats]{setNames}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{reexports}}
 \code{\link[utils]{type.convert}}
}
