% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_iris_source.R
\name{import_iris_source}
\alias{import_iris_source}
\title{Load IRIS Source into toxval_source}
\usage{
import_iris_source(
  db,
  infile1 = "IRIS_non_cancer_clean 2022-10-21.xlsx",
  infile2 = "IRIS_cancer_clean 2022-10-21.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile1}{The input file ./iris/iris_files/IRIS_non_cancer_clean 2020-05-27.xlsx}

\item{infile2}{The input file ./iris/iris_files/IRIS_cancer_clean 2020-05-27.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load IRIS Source into toxval_source
}
