% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_lanl_source.R
\name{import_lanl_source}
\alias{import_lanl_source}
\title{Load lanl Source into dev_toxval_source_v2.}
\usage{
import_lanl_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./lanl/lanl_files/ESLs_R3.3.xlsx}
}
\description{
Load lanl Source into dev_toxval_source_v2.
}
