% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_niosh_source.R
\name{import_niosh_source}
\alias{import_niosh_source}
\title{Load niosh Source into dev_toxval_source_v4.}
\usage{
import_niosh_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./niosh/niosh_files/niosh_IDLH_2020.xlsx}
}
\description{
Load niosh Source into dev_toxval_source_v4.
}
