% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_niosh_source.R
\name{import_niosh_source}
\alias{import_niosh_source}
\title{FUNCTION_TITLE}
\usage{
import_niosh_source(db, infile = "niosh_IDLH_2020.xlsx", chem.check.halt = T)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./niosh/niosh_files/niosh_IDLH_2020.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load NIOSH Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
}
