% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_niosh_source.R
\name{import_niosh_source}
\alias{import_niosh_source}
\title{import_niosh_source}
\usage{
import_niosh_source(
  db,
  chem.check.halt = FALSE,
  do.reset = FALSE,
  do.insert = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}

\item{do.reset}{If TRUE, delete data from the database for this source before}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}
}
\value{
None; data is pushed to toxval_source
}
\description{
Load NIOSH data into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}
 \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_extract}}
 \code{\link[tidyr]{separate}}, \code{\link[tidyr]{drop_na}} \code{\link[tidyr]{pivot_longer}}
}
