% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_opp_source.R
\name{import_opp_source}
\alias{import_opp_source}
\title{Load opp Source into dev_toxval_source_v2.}
\usage{
import_opp_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./opp/opp_files/OPP RfD.xlsx}
}
\description{
Load opp Source into dev_toxval_source_v2.
}
