% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_opp_source.R
\name{import_opp_source}
\alias{import_opp_source}
\title{FUNCTION_TITLE}
\usage{
import_opp_source(
  db,
  infile = "source_opp_raw_20230131.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./opp/opp_files/OPP RfD.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load OPP Source into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}
 \code{\link[stringr]{str_trim}}
}
