% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_opp_source.R
\name{import_opp_source}
\alias{import_opp_source}
\title{Load OPP Source into toxval_source}
\usage{
import_opp_source(db, infile = "OPP RfD.xlsx", chem.check.halt = T)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile}{The input file ./opp/opp_files/OPP RfD.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load OPP Source into toxval_source
}
