% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_oppt_source.R
\name{import_oppt_source}
\alias{import_oppt_source}
\title{Load oppt Source Info into dev_toxval_source_v2.}
\usage{
import_oppt_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{infile}{The input file ./oppt/oppt_files/OPPT_data_20181219.xlsx}
}
\description{
Load oppt Source Info into dev_toxval_source_v2.
}
