% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_penn_source.R
\name{import_penn_source}
\alias{import_penn_source}
\title{Load penn Source into dev_toxval_source_v2.}
\usage{
import_penn_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./penn/penn_files/Penn DEP Table 5a.xlsx}
}
\description{
Load penn Source into dev_toxval_source_v2.
}
