% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pfas_150_sem_source.R
\name{import_pfas_150_sem_source}
\alias{import_pfas_150_sem_source}
\title{Load PFAS 150 SEM Source data into dev_toxval_source_v2.}
\usage{
import_pfas_150_sem_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{infile}{The input file ./PFAS 150 SEM/PFAS 150 SEM_files/PFAS150 animal study template combined_clearance with DTXSID and CASRN.xlsx}
}
\description{
Load PFAS 150 SEM Source data into dev_toxval_source_v2.
}
