% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pfas_150_sem_v2_source.R
\name{import_pfas_150_sem_v2_source}
\alias{import_pfas_150_sem_v2_source}
\title{Load PFAS 150 SEM V2 Source data into toxval_source}
\usage{
import_pfas_150_sem_v2_source(db, chem.check.halt = F)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{chem.check.halt}{If TRUE and there are problems with chemicals CASRN checks, halt the program}
}
\description{
Load PFAS 150 SEM V2 Source data into toxval_source
}
