% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pfas_summary_pods_source.R
\name{import_pfas_summary_pods_source}
\alias{import_pfas_summary_pods_source}
\title{Load PFAS Summary PODs into toxval_source}
\usage{
import_pfas_summary_pods_source(
  db,
  infile1 = "PFAS 150 Study Level PODs_061920.xlsx",
  infile2 = "CompToxChemicalsDashboard-Batch-Search_2020-07-20_17_18_42.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{infile1}{The input file ./PFAS Summary PODs/PFAS Summary PODs_files/PFAS 150 Study Level PODs_061920.xlsx}

\item{infile2}{The input file ./PFAS Summary PODs/PFAS Summary PODs_files/CompToxChemicalsDashboard-Batch-Search_2020-07-20_17_18_42.xls}

\item{chem.check.halt}{If TRUE and there are problems with chemicals CASRN checks, halt the program}
}
\description{
Load PFAS Summary PODs into toxval_source
}
