% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pprtv_ncea_source.R
\name{import_pprtv_ncea_source}
\alias{import_pprtv_ncea_source}
\title{Load PPRTV NCEA Source Info into toxval_source}
\usage{
import_pprtv_ncea_source(
  db,
  csvfile = "../pprtv_ncea/pprtv_ncea_files/dose_reg2.csv",
  scrapepath = "../pprtv_ncea/PPRTV_scrape2020-04-08.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{csvfile}{The input csv file ./pprtv_ncea/pprtv_ncea_files/dose_reg2.csv}

\item{scrapepath}{The path for new_pprtv_ncea_scrape_table file ./pprtv_ncea/PPRTV_scrape2020-04-08.xlsx}

\item{chem.check.halt}{If TRUE and there are problems with chemicals CASRN checks, halt the program}
}
\description{
Load PPRTV NCEA Source Info into toxval_source
}
