% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pprtv_ornl_source.R
\name{import_pprtv_ornl_source}
\alias{import_pprtv_ornl_source}
\title{Load pprtv_ornl Source into dev_toxval_source_v2.}
\usage{
import_pprtv_ornl_source(toxval.db, infile)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./pprtv_ornl/pprtv_ornl_files/new_PPRTV_ORNL cancer noncancer.xlsx}
}
\description{
Load pprtv_ornl Source into dev_toxval_source_v2.
}
