% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pprtv_ornl_source.R
\name{import_pprtv_ornl_source}
\alias{import_pprtv_ornl_source}
\title{Load PPRTV ORNL Source into toxval_source}
\usage{
import_pprtv_ornl_source(
  db,
  infile = "new_PPRTV_ORNL cancer noncancer.xlsx",
  chem.check.halt = F
)
}
\arguments{
\item{db}{The version of toxval into which the source is loaded.}

\item{infile}{The input file ./pprtv_ornl/pprtv_ornl_files/new_PPRTV_ORNL cancer noncancer.xlsx}

\item{chem.check.halt}{If TRUE and there are problems with chemicals CASRN checks, halt the program}
}
\description{
Load PPRTV ORNL Source into toxval_source
}
