% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rsl_source.R
\name{import_rsl_source}
\alias{import_rsl_source}
\title{Load RSL Source Info into toxval source database}
\usage{
import_rsl_source(
  db,
  infile1a = "final_rsl_thq_combined_nov21.xlsx",
  infile1b = "final_rsl_subchronic_nov21.xlsx",
  infile2 = "general_info_nov_21.xlsx",
  infile3 = "key_description_nov_21.xlsx",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval into which the source info is loaded.}

\item{infile1a}{The input file ./rsl/rsl_files/final_rsl_thq_combined_nov21.xlsx}

\item{infile1b}{The input file ./rsl/rsl_files/final_rsl_subchronic_nov21.xlsx}

\item{infile2}{The input file ./rsl/rsl_files/general_info_nov_21.xlsx}

\item{infile3}{The input file ./rsl/rsl_files/key_description_nov_21.xlsx}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\description{
Load RSL Source Info into toxval source database
}
