% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_source_iuclid.R
\name{import_source_iuclid}
\alias{import_source_iuclid}
\title{A generic template for adding data to toxval_source for a new source}
\usage{
import_source_iuclid(db, subf, chem.check.halt = F)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{subf}{The subfolder containing the IUCLID subsource}

\item{chem.check.halt}{If TRUE and there are bad chemical names or casrn,}
}
\description{
A generic template for adding data to toxval_source for a new source
}
