% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_test_source.R
\name{import_test_source}
\alias{import_test_source}
\title{Load test Source data into dev_toxval_source_v4.}
\usage{
import_test_source(toxval.db, infile1, infile2)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{infile1}{The input file ./test/test_files/TEST data.xlsx}

\item{infile2}{The input file ./test/test_files/test_chemicals_invitrodb.csv to map casrn to names from prod_internal_invitrodb_v3_2.chemical}
}
\description{
Load test Source data into dev_toxval_source_v4.
}
