% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_test_source.R
\name{import_test_source}
\alias{import_test_source}
\title{FUNCTION_TITLE}
\usage{
import_test_source(
  db,
  infile1 = "TEST data.xlsx",
  infile2 = "test_chemicals_invitrodb.csv",
  chem.check.halt = T
)
}
\arguments{
\item{db}{The version of toxval_source into which the source info is loaded.}

\item{infile1}{The input file ./test/test_files/TEST data.xlsx}

\item{infile2}{The input file ./test/test_files/test_chemicals_invitrodb.csv to map casrn to names from prod_internal_invitrodb_v3_2.chemical}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Load TEST Source data into toxval_source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[openxlsx]{read.xlsx}}
 \code{\link[utils]{read.table}}
 \code{\link[stringr]{str_count}}
}
