% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.audit.table.R
\name{init.audit.table}
\alias{init.audit.table}
\title{init.audit.table}
\usage{
init.audit.table(db, do.halt = FALSE, verbose = FALSE)
}
\arguments{
\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}

\item{s_tbl}{Source table name to apply changes to}

\item{field_list}{List of current field names in source table}
}
\value{
None
}
\description{
Create audit table and add BEFORE UPDATE audit triggers to source_* tables

Create audit table and add BEFORE UPDATE audit triggers to source_* tables
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
