% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS_cancer_clean.R
\name{iris.cancer.clean}
\alias{iris.cancer.clean}
\title{create IRIS cancer clean source file from https://cfpub.epa.gov/ncea/iris/search/index.cfm
by selecting cancer, oral and inhalation data, and toxicity value checkboxes.}
\usage{
iris.cancer.clean(infile)
}
\arguments{
\item{infile}{The input file ../iris/iris_files/excelView.xls}
}
\description{
create IRIS cancer clean source file from https://cfpub.epa.gov/ncea/iris/search/index.cfm
by selecting cancer, oral and inhalation data, and toxicity value checkboxes.
}
