% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_message.R
\name{log_message}
\alias{log_message}
\title{Function to combine output log with output message}
\usage{
log_message(log_df, message_df_col)
}
\arguments{
\item{log_df}{Dataframe to which the log information will be appended}

\item{message_df_col}{New message to add}
}
\description{
Function to combine output log with output message
}
