% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sql_file.R
\name{parse_sql_file}
\alias{parse_sql_file}
\title{parse_sql_file}
\usage{
parse_sql_file(filepath = NULL)
}
\arguments{
\item{filepath}{Input SQL filepath}
}
\value{
SQL query strings
}
\description{
Function to parse SQL file into SQL query strings
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readr]{read_lines}}
 \code{\link[stringr]{str_trim}}
}
