% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_prescreening_summary.R
\name{qc_prescreening_summary}
\alias{qc_prescreening_summary}
\title{Runs a database query and returns a result set}
\usage{
qc_prescreening_summary(src_tbl = NULL, outputDir = NULL, db = NULL)
}
\arguments{
\item{src_tbl}{a toxval source table name}

\item{outputDir}{optional directory path to save output file in}

\item{db}{the name of the database}
}
\description{
Runs a database query and returns a result set
}
