% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQuery_psql.R
\name{runQuery_psql}
\alias{runQuery_psql}
\title{runQuery_psql}
\usage{
runQuery_psql(query, db, do.halt = T, verbose = T)
}
\arguments{
\item{query}{a properly formatted SQL query as a string}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}
}
\value{
Query results
}
\description{
Runs a PSQL database query and returns a result set
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[RMySQL]{character(0)}}
 \code{\link[RPostgreSQL]{PostgreSQL}}
 \code{\link[DBI]{dbGetQuery}}
}
