% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runUpdate.R
\name{runUpdate}
\alias{runUpdate}
\title{FUNCTION_TITLE}
\usage{
runUpdate(
  table,
  updateQuery = NULL,
  updated_df = NULL,
  db,
  do.halt = TRUE,
  verbose = FALSE,
  trigger_check = TRUE
)
}
\arguments{
\item{table}{PARAM_DESCRIPTION}

\item{updateQuery}{a properly formatted SQL query as a string in the form of an UPDATE INNER JOIN}

\item{updated_df}{a dataframe of updated data to temporarily write to database for INNER JOIN}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}

\item{trigger_check}{if FALSE, audit triggers are ignored/bypassed}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Runs a database query and returns a result set
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[RMySQL]{character(0)}}, \code{\link[RMySQL]{MySQLDriver-class}}
 \code{\link[DBI]{dbSendStatement}}
}
