% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDBConn.R
\name{setDBConn}
\alias{setDBConn}
\title{set SQL connection to the database}
\usage{
setDBConn(server, user, password)
}
\arguments{
\item{server}{SQL server on which relevant database lives}

\item{user}{SQL username to access database}

\item{password}{SQL password corresponding to username}
}
\description{
set SQL connection to the database
}
