% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDBConn.R
\name{setDBConn}
\alias{setDBConn}
\title{FUNCTION_TITLE}
\usage{
setDBConn(server = "ccte-mysql-res.epa.gov", user, password)
}
\arguments{
\item{server}{SQL server on which relevant database lives}

\item{user}{SQL username to access database}

\item{password}{SQL password corresponding to username}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
set SQL connection to the database
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
