% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPSQLDBConn.R
\name{setPSQLDBConn}
\alias{setPSQLDBConn}
\title{setPSQLDBConn}
\usage{
setPSQLDBConn(server, port, user, password, api_auth)
}
\arguments{
\item{server}{SQL server on which relevant database lives}

\item{port}{port to use}

\item{user}{SQL username to access database}

\item{password}{SQL password corresponding to username}

\item{api_auth}{API Key for CCTE API's}
}
\value{
None
}
\description{
set PSQL connection to the database
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
