% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPSQLDBConn.R
\name{setPSQLDBConn}
\alias{setPSQLDBConn}
\title{set PSQL connection to the database}
\usage{
setPSQLDBConn(server, port, user, password)
}
\arguments{
\item{server}{SQL server on which relevant database lives}

\item{user}{SQL username to access database}

\item{password}{SQL password corresponding to username}
}
\description{
set PSQL connection to the database
}
