% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_chemical.extra.R
\name{source_chemical.extra}
\alias{source_chemical.extra}
\title{special process to deal with source chemicals for extra source (cancer, genetox, skin_eye, etc)}
\usage{
source_chemical.extra(
  toxval.db,
  source.db,
  res,
  source,
  chem.check.halt = FALSE,
  casrn.col = "casrn",
  name.col = "name",
  verbose = F
)
}
\arguments{
\item{toxval.db}{The version of toxval into which the source info is loaded.}

\item{source.db}{The source database version}

\item{source}{The source to be processed #' @param chem.check.halt If TRUE, halt if there are errors in the chemical checking}

\item{casrn.col}{Name of the column containing the CASRN}

\item{name.col}{Name of the column containing chemical names}

\item{verbose}{If TRUE, output extra diagnostics information}
}
\description{
special process to deal with source chemicals for extra source (cancer, genetox, skin_eye, etc)
}
