% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_chemical.process.R
\name{source_chemical.process}
\alias{source_chemical.process}
\title{FUNCTION_TITLE}
\usage{
source_chemical.process(
  db,
  res,
  source,
  table,
  chem.check.halt = FALSE,
  casrn.col = "casrn",
  name.col = "name",
  verbose = FALSE
)
}
\arguments{
\item{db}{The version of toxval into which the source info is loaded.}

\item{res}{The input dataframe to which chemical information will be added}

\item{source}{The source to process}

\item{table}{Name of the database table}

\item{chem.check.halt}{If TRUE, stop if there are problems with the chemical mapping}

\item{casrn.col}{The name of the column containing the CASRN}

\item{name.col}{The name ofhte column containing hte chemical name}

\item{verbose}{If TRUE, write out diagnostic messages #'}
}
\value{
Returns the original dataframe with a chemical_id appended
}
\description{
Deal with the process of making the source_chemical information
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{unite}}
 \code{\link[utils]{head}}
 \code{\link[digest]{digest}}
}
