% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_prep_and_load.R
\name{source_prep_and_load}
\alias{source_prep_and_load}
\title{Prep the source data aand load}
\usage{
source_prep_and_load(
  db,
  source,
  table,
  res,
  do.reset = FALSE,
  do.insert = FALSE,
  chem.check.halt = FALSE
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{source}{Name of the source}

\item{table}{Name of the database table}

\item{res}{The data frame to be processed}

\item{do.reset}{If TRUE, delete data from the database for this source before
inserting new data. Default FALSE}

\item{do.insert}{If TRUE, insert data into the database, default TRUE}

\item{chem.check.halt}{If TRUE, stop the execution if there are errors in the
chemical  mapping}
}
\description{
Prep the source data aand load
}
