% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_prep_and_load.R
\name{source_prep_and_load}
\alias{source_prep_and_load}
\title{source_prep_and_load}
\usage{
source_prep_and_load(
  db,
  source,
  table,
  res,
  do.reset = FALSE,
  do.insert = FALSE,
  chem.check.halt = FALSE,
  verbose = FALSE,
  hashing_cols = NULL
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{source}{Name of the source}

\item{table}{Name of the database table}

\item{res}{The data frame to be processed}

\item{do.reset}{If TRUE, delete data from the database for this source before #' inserting new data. Default FALSE}

\item{do.insert}{If TRUE, insert data into the database, default FALSE}

\item{chem.check.halt}{If TRUE, stop the execution if there are errors in the #' chemical mapping}

\item{verbose}{If TRUE, write out diagnostic messages #'}

\item{hashing_cols}{Optional list of columns to use for generating source_hash}
}
\value{
None
}
\description{
Prep the source data aand load
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tidyr]{reexports}}
}
