% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_set_defaults.R
\name{source_set_defaults}
\alias{source_set_defaults}
\title{FUNCTION_TITLE}
\usage{
source_set_defaults(res, source)
}
\arguments{
\item{res}{The input dataframe}

\item{source}{The data source name}
}
\value{
Returns the input dataframe with defaults set
}
\description{
Set default value for NAs - just set NA to "-" for columns of type character
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
