% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.all.R
\name{toxval.load.all}
\alias{toxval.load.all}
\title{Load and process all information into ToxValDB. The entire process can be run with
one command: toxval.load.all(toxval.db=...,source.db=..., do.all=T)
It can also be run in stages, but needs to be run in the order of the do.X parameters
listed here. If any earlier step is run, all of the subsequent steps need to be rerun.}
\usage{
toxval.load.all(
  toxval.db,
  source.db,
  log = F,
  do.init = F,
  do.reset = F,
  do.load = F,
  do.post = F
)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source database from which information is pulled.}

\item{log}{If TRUE write the output from each load script to a log file}

\item{do.init}{If True, clean out all of the database tables}

\item{do.reset}{If TRUE, empty the database to restart}

\item{do.load}{If TRUE, load all of the source}
}
\description{
Load and process all information into ToxValDB. The entire process can be run with
one command: toxval.load.all(toxval.db=...,source.db=..., do.all=T)
It can also be run in stages, but needs to be run in the order of the do.X parameters
listed here. If any earlier step is run, all of the subsequent steps need to be rerun.
}
