% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.atsdr.pfas.2021.R
\name{toxval.load.atsdr.pfas.2021}
\alias{toxval.load.atsdr.pfas.2021}
\title{Load data ATSDR PFAS 2021 data from toxval_source to toxval}
\usage{
toxval.load.atsdr.pfas.2021(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{log}{If TRUE, send messages to a log file}
}
\description{
Load data ATSDR PFAS 2021 data from toxval_source to toxval
}
