% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.atsdr.pfas.R
\name{toxval.load.atsdr.pfas}
\alias{toxval.load.atsdr.pfas}
\title{Load new_atsdr_pfas from toxval_source to toxval}
\usage{
toxval.load.atsdr.pfas(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load new_atsdr_pfas from toxval_source to toxval
}
