% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.bcfbaf.R
\name{toxval.load.bcfbaf}
\alias{toxval.load.bcfbaf}
\title{Load the Arnot BAF / BCF data}
\usage{
toxval.load.bcfbaf(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The database to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load the Arnot BAF / BCF data
}
