% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.cosmos.R
\name{toxval.load.cosmos}
\alias{toxval.load.cosmos}
\title{Load cosmos from source to toxval}
\usage{
toxval.load.cosmos(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{verbose}{If TRUE, print out extra diagnostic messages}
}
\description{
Load cosmos from source to toxval
}
