% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.dod.ered.R
\name{toxval.load.dod.ered}
\alias{toxval.load.dod.ered}
\title{Load the DOD ERED data from toxval_source to toxval}
\usage{
toxval.load.dod.ered(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load the DOD ERED data from toxval_source to toxval
}
