% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.echa.echemportal.api.R
\name{toxval.load.echa.echemportal.api}
\alias{toxval.load.echa.echemportal.api}
\title{Load ECHA eChemPortal API data from toxval_source to toxval}
\usage{
toxval.load.echa.echemportal.api(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source database to use.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load ECHA eChemPortal API data from toxval_source to toxval
}
