% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.ecotox.R
\name{toxval.load.ecotox}
\alias{toxval.load.ecotox}
\title{Load ECOTOX from toxval_source to toxval}
\usage{
toxval.load.ecotox(toxval.db, verbose = T, do.load = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{verbose}{Whether the loaded rows should be printed to the console.}

\item{do.load}{If TRUE, load the data from the input file and put into a global variable}
}
\description{
Load ECOTOX from toxval_source to toxval
}
