% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.efsa2.R
\name{toxval.load.efsa2}
\alias{toxval.load.efsa2}
\title{Load new_efsa2 from toxval_source to toxval}
\usage{
toxval.load.efsa2(toxval.db, source.db, verbose = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source databse from which data should be loaded}

\item{verbose}{Whether the loaded rows should be printed to the console.}
}
\description{
Load new_efsa2 from toxval_source to toxval
}
