% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.epa_aegl.R
\name{toxval.load.epa_aegl}
\alias{toxval.load.epa_aegl}
\title{Load the epa_aegl (old ACToR - flex)data  from toxval sourcedb to toxval}
\usage{
toxval.load.epa_aegl(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{verbose}{If TRUE, output extra debug information}
}
\description{
Load the epa_aegl (old ACToR - flex)data  from toxval sourcedb to toxval
}
