% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.pprtv.ncea.R
\name{toxval.load.pprtv.ncea}
\alias{toxval.load.pprtv.ncea}
\title{Load PPRTV (NCEA) from toxval source to toxval}
\usage{
toxval.load.pprtv.ncea(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The version of toxval_source from which the tables are loaded.}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load PPRTV (NCEA) from toxval source to toxval
}
