% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.pprtv.ornl.R
\name{toxval.load.pprtv.ornl}
\alias{toxval.load.pprtv.ornl}
\title{Load PPRTV (ORNL) from toxval_source to toxval}
\usage{
toxval.load.pprtv.ornl(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{source.db}{The source databse from which data should be loaded}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load PPRTV (ORNL) from toxval_source to toxval
}
