% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.rsl.R
\name{toxval.load.rsl}
\alias{toxval.load.rsl}
\title{Load the RSL data - the source database needs to be updated periodically}
\usage{
toxval.load.rsl(toxval.db, source.db, verbose = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{verbose}{If TRUE, output extra debug information}
}
\description{
Load the RSL data - the source database needs to be updated periodically
}
