% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.toxrefdb3.R
\name{toxval.load.toxrefdb3}
\alias{toxval.load.toxrefdb3}
\title{Load ToxRefdb data to toxval}
\usage{
toxval.load.toxrefdb3(toxval.db, verbose = F, do.init = T)
}
\arguments{
\item{toxval.db}{The version of toxval into which the tables are loaded.}

\item{verbose}{Whether the loaded rows should be printed to the console.}

\item{do.init}{if TRUE, read the data in from the toxrefdb database and set up the matrix}
}
\description{
Load ToxRefdb data to toxval
}
