% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.load.usgs_hbsl.R
\name{toxval.load.usgs_hbsl}
\alias{toxval.load.usgs_hbsl}
\title{Load the usgs_hbsl (old ACToR - flex)data  from toxval source db to toxval}
\usage{
toxval.load.usgs_hbsl(toxval.db, source.db, log = F)
}
\arguments{
\item{toxval.db}{The database version to use}

\item{source.db}{The source database}

\item{log}{If TRUE, send output to a log file}
}
\description{
Load the usgs_hbsl (old ACToR - flex)data  from toxval source db to toxval
}
