% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_hash_vectorized.R,
%   R/toxval_source.hash.and.load.R
\name{source_hash_vectorized}
\alias{source_hash_vectorized}
\alias{toxval_source.hash.and.load}
\title{source_hash_vectorized}
\usage{
source_hash_vectorized(res, hashing_cols)

toxval_source.hash.and.load(
  db = "dev_toxval_source_v5",
  source,
  table,
  do.reset = FALSE,
  do.insert = FALSE,
  res,
  hashing_cols = NULL
)
}
\arguments{
\item{res}{The data frame to be processed}

\item{hashing_cols}{Optional list of columns to use for generating source_hash}

\item{db}{The version of toxval_source into which the source is loaded.}

\item{source}{Name of the source}

\item{table}{Name of the database table}

\item{do.reset}{If TRUE, delete data from the database for this source before #' inserting new data. Default FALSE}

\item{do.insert}{If TRUE, insert data into the database, default False}
}
\value{
Input dataframe with new source_hash field

OUTPUT_DESCRIPTION
}
\description{
Generate the hash key for a source table based on hashing columns

Add the hash key to the source tables and add the new rows
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[digest]{digest}}
 \code{\link[dplyr]{distinct}}

\code{\link[digest]{digest}}
 \code{\link[dplyr]{distinct}}
}
