% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval_source.hash.and.load.R
\name{toxval_source.hash.and.load}
\alias{toxval_source.hash.and.load}
\title{FUNCTION_TITLE}
\usage{
toxval_source.hash.and.load(
  db = "dev_toxval_source_v5",
  source,
  table,
  do.reset = FALSE,
  do.insert = FALSE,
  res
)
}
\arguments{
\item{db}{The version of toxval_source into which the source is loaded.}

\item{source}{Name of the source}

\item{table}{Name of the database table}

\item{do.reset}{If TRUE, delete data from the database for this source before #' inserting new data. Default FALSE}

\item{do.insert}{If TRUE, insert data into the database, default False}

\item{res}{The data frame to be processed}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Add the hash key to the source tables and add the new rows
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[digest]{digest}}
 \code{\link[dplyr]{distinct}}
}
