% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_chemical_preferred_info_by_dtxsid.R
\name{update_chemical_preferred_info_by_dtxsid}
\alias{update_chemical_preferred_info_by_dtxsid}
\title{Set the name and casrn in the source_chemical table based on CCTE API}
\usage{
update_chemical_preferred_info_by_dtxsid(source.db)
}
\arguments{
\item{source.db}{The database version to use}
}
\description{
Set the name and casrn in the source_chemical table based on CCTE API
}
